### 2023-01-12 #68 @lwasser
hi @NickleDave responding here so it's visible that this issue is being worked on! I am looking for an editor for this package now given I can't step in  due to COI! :) but please know i've reached out to a few people and am in the process of finding an editor.
