### 2023-01-13 #68 @NickleDave
Hi @cmarmo! Nice to meet you as well.

Looking forward to working with you on the review, I am sure I will learn a lot.

A week is perfect, I should have survived a deadline at my day job by then :laughing:
