### 2023-01-15 #68 @cmarmo
Hi @NickleDave, below are the basic checks that your package needs to pass
to begin our review.

  - [x] The package imports properly into a standard Python environment `import package-name`.
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
    - The `contributing.md` file is located in the `.github/` folder, while pyOpenSci [recommends](https://www.pyopensci.org/python-package-guide/documentation/contributing.html) including it in the root repository.
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---

The package is already in a very good shape, thank you for all your work!

I would like to mention that the `contributing.md` file is located in the `.github/` directory and not in the root repository as [recommended](https://www.pyopensci.org/python-package-guide/documentation/contributing.html)  by the pyOpenSci documentation.
I don't have any strong opinion about that, but I must confess that I had to look for it for a moment before remembering that it could be placed in the `.github` directory. I guess having it in the root repo will make it more discoverable directly on-line (without downloading the sources).
Feel free to move it or not, depending on your point of view.

I'm going to look for reviewers for your package now.
I hope this will not take too much time... I will let you know anyway by the end of the week.
Thanks for your patience!
