### 2023-01-15 #68 @NickleDave
:pray: thank you @cmarmo

> the contributing.md file is located in the .github/ directory and not in the root repository as [recommended](https://www.pyopensci.org/python-package-guide/documentation/contributing.html) by the pyOpenSci documentation.
> I guess having it in the root repo will make it more discoverable directly on-line

I can move it to the root, and I agree with you that it's easier to spot that way for someone looking at the repo e.g. on GitHub in the browser.

I think I was basically copying what I saw in [numpy](https://github.com/numpy/numpy/blob/main/.github/CONTRIBUTING.md)--I thought the file had to be there for the link to show up when someone opened and issue.

But now I see the [GitHub docs](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) say it can be there, in the root, or docs. Good to know.

So I'll just move it.
