### 2023-01-15 9d435f6
DOC: Rename .github/contributing.md ./CONTRIBUTING.md [skip ci]
