### 2023-01-23 #68 @shaupert
Hi  @cmarmo, @NickleDave, and @rhine3,
Thank you for inviting me to review the crowsetta package. It looks great.
I'm a novice in building Python packages but I'll try to do my best to review the package.
I'm very exciting to test crowsetta and see you I can use it :)
