### 2023-01-23 #68 @lwasser
@cmarmo this is all awesome! @rhine3 and  @shaupert  welcome to pyopensci!!

@shaupert if you have any questions at all about the review process please get in touch. It is useful for us to have a review that focuses on usability, documentation, etc. You are welcome to reach out to us with questions at any time - so please feel free to do that if you wish / if it would be helpful.
