### 2023-01-28 bdc578b
Add remove-silent-labels-textgrid to toc in howto.md
