### 2023-01-28 e753797
DEV/DOC: enable myst extension 'dollarmath' in conf.py
