### 2023-01-28 e43e360
TST: Add unit tests for annot2df, GenericSeq.to_df
