### 2023-01-28 #68 @NickleDave
Thank you @shaupert and @rhine3. Just a heads up that I just now merged a branch I had in process before submission, that mostly cleaned up + added docs--promise not to do any further dev for now. Looking forward to your reviews.
