### 2023-01-29 c2f657d
DEV/DOC: Add missing 'doc' dependency osfclient
