+++ pyproject.toml
@@ -46,6 +46,7 @@ doc = [
     "jupytext >=1.13.8",
     "librosa >=0.9.1",
     "myst-nb >=0.15.0",
+    "osfclient >=0.0.5",
     "pyprojroot >=0.2.0",
     "seaborn >=0.11.2",
     "Sphinx >=3.4.3",
