### 2023-01-29 248b0a2
DOC: Update CHANGELOG after merging #216 [skip ci]
