### 2023-01-29 7c22a01
DOC/CLN: Revise + clean up index.md and 2 how-tos
- Make minor revisions/typo fixes etc in index.md
- Revise, add headers in doc/howto/remove-silent-labels-textgrid.md
- Clean up doc/howto/convert-generic-seq.md
- Make fixes in doc/howto/remove-silent-labels-textgrid.md
- Add timeout to yaml front matter in doc/tutorial.md
