### 2023-02-09 vocalpy/crowsetta#218 @rhine3
BUG: tests in test_formats/test_seq/test_generic.py failed
#### Before submitting a bug, please make sure the issue hasn't been already addressed by searching through [the past issues](https://github.com/NickleDave/vak/issues)

**Describe the bug**
While running tests with Pytest I found that 6 of the tests in `test_formats/test_seq/test_generic.py` failed on my machine. The short test summary info is below; longer test summaries are copied at the end of this issue.
```
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_s_only - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_s_and_ind - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_sample_only - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_annot2df_onset_offset_s_only - AttributeError: 'GenericSeq' object has no attribute 'to_df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_to_df_onset_offset_s_and_ind - AttributeError: 'GenericSeq' object has no attribute 'to_df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_to_df_onset_offset_sample_only - AttributeError: 'GenericSeq' object has no attribute 'to_df'
```

**To Reproduce**
I cloned the GitHub repository on my local machine. I installed `crowsetta` in a Python 3.10.9 Anaconda environment and also conda installed `pytest`. I navigated to the `tests` directory and ran the command `pytest *`.

**Expected behavior**
I expected the tests to pass! :)

**Screenshots**
Errors are reproduced in text below instead

 - Operating System: Mac
 - Version [e.g. 22]. You can determine this in Python by running `vak.__version__` - My `crowsetta` version is '4.0.0.post2'; by the way, you may want to update the sample issue text here to say "You can determine this in Python by importing `crowsetta` and running `crowsetta.__version__`"

**Additional context**
I created this issue while performing the [PyOpenSci review of this package](https://github.com/pyOpenSci/software-submission/issues/68). Full text of tests is reproduced below.

```
=================================================================================================================================== FAILURES ===================================================================================================================================
____________________________________________________________________________________________________________ TestAnnot2DfFunction.test_annot2df_onset_offset_s_only ____________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestAnnot2DfFunction object at 0x7f8049738bb0>
notmat_paths = [PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/cbins/gy6or6/032312/gy6...iew/crowsetta/tests/fixtures/../data_for_tests/cbins/gy6or6/032312/gy6or6_baseline_230312_0816.179.cbin.not.mat'), ...]
notmat_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/notmat_gy6or6_032312.csv')

    def test_annot2df_onset_offset_s_only(self,
                                          notmat_paths,
                                          notmat_as_generic_seq_csv):
        """test whether `annot2df` works when
        the annotations have onsets and offsets specified in seconds only.
        To test this we use the 'notmat' format.
        """
        annot_list = [crowsetta.formats.seq.NotMat.from_file(notmat_path).to_annot()
                      for notmat_path in notmat_paths]
        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = crowsetta.formats.seq.generic.annot2df(annot_list,
                                                            basename=True)
E       AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'

test_formats/test_seq/test_generic.py:101: AttributeError
__________________________________________________________________________________________________________ TestAnnot2DfFunction.test_annot2df_onset_offset_s_and_ind ___________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestAnnot2DfFunction object at 0x7f8049738ac0>, birdsong_rec_xml_file = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/birdsongrec/Bird0/Annotation.xml')
birdsong_rec_wav_path = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/birdsongrec/Bird0/Wave')
birdsongrec_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/birdsongrec_Bird0_Annotation.csv')

    def test_annot2df_onset_offset_s_and_ind(self,
                                             birdsong_rec_xml_file,
                                             birdsong_rec_wav_path,
                                             birdsongrec_as_generic_seq_csv):
        """test whether `annot2df` works when
        the annotations have onsets and offsets specified in seconds and in samples.
        To test this we use the 'birdsong-recognition-dataset' format.
        """
        birdsongrec = crowsetta.formats.seq.BirdsongRec.from_file(annot_path=birdsong_rec_xml_file,
                                                                  wav_path=birdsong_rec_wav_path,
                                                                  concat_seqs_into_songs=True)
        annot_list = birdsongrec.to_annot()
        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = crowsetta.formats.seq.generic.annot2df(annot_list,
                                                            basename=True)
E       AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'

test_formats/test_seq/test_generic.py:123: AttributeError
_________________________________________________________________________________________________________ TestAnnot2DfFunction.test_annot2df_onset_offset_sample_only __________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestAnnot2DfFunction object at 0x7f8049739f30>
kaggle_phn_paths = [PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/timit_kaggle/dr1-fvmh0/...'/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/timit_kaggle/dr1-fvmh0/si836.phn')]
timit_phn_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/timit-dr1-fvmh0-phn.csv')

    def test_annot2df_onset_offset_sample_only(self,
                                               kaggle_phn_paths,
                                               timit_phn_as_generic_seq_csv):
        """test whether `annot2df` works when
        the annotations have onsets and offsets specified in seconds only.
        To test this we use the 'timit' format."""
        annot_list = [crowsetta.formats.seq.Timit.from_file(phn_path).to_annot()
                      for phn_path in kaggle_phn_paths]

        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = crowsetta.formats.seq.generic.annot2df(annot_list,
                                                            basename=True)
E       AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'

test_formats/test_seq/test_generic.py:144: AttributeError
____________________________________________________________________________________________________________ TestGenericSeqClass.test_annot2df_onset_offset_s_only _____________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestGenericSeqClass object at 0x7f8049e0aa10>
notmat_paths = [PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/cbins/gy6or6/032312/gy6...iew/crowsetta/tests/fixtures/../data_for_tests/cbins/gy6or6/032312/gy6or6_baseline_230312_0816.179.cbin.not.mat'), ...]
notmat_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/notmat_gy6or6_032312.csv')

    def test_annot2df_onset_offset_s_only(self,
                                          notmat_paths,
                                          notmat_as_generic_seq_csv):
        """test whether `annot2df` works when
        the annotations have onsets and offsets specified in seconds only.
        To test this we use the 'notmat' format.
        """
        annot_list = [crowsetta.formats.seq.NotMat.from_file(notmat_path).to_annot()
                      for notmat_path in notmat_paths]
        generic_seq = crowsetta.formats.seq.GenericSeq(annots=annot_list)
        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = generic_seq.to_df(basename=True)
E       AttributeError: 'GenericSeq' object has no attribute 'to_df'

test_formats/test_seq/test_generic.py:479: AttributeError
____________________________________________________________________________________________________________ TestGenericSeqClass.test_to_df_onset_offset_s_and_ind _____________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestGenericSeqClass object at 0x7f8049e0a890>, birdsong_rec_xml_file = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/birdsongrec/Bird0/Annotation.xml')
birdsong_rec_wav_path = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/birdsongrec/Bird0/Wave')
birdsongrec_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/birdsongrec_Bird0_Annotation.csv')

    def test_to_df_onset_offset_s_and_ind(self,
                                          birdsong_rec_xml_file,
                                          birdsong_rec_wav_path,
                                          birdsongrec_as_generic_seq_csv):
        """test whether `GenericSeq.to_df` works when
        the annotations have onsets and offsets specified in seconds and in samples.
        To test this we use the 'birdsong-recognition-dataset' format.
        """
        birdsongrec = crowsetta.formats.seq.BirdsongRec.from_file(annot_path=birdsong_rec_xml_file,
                                                                  wav_path=birdsong_rec_wav_path,
                                                                  concat_seqs_into_songs=True)
        annot_list = birdsongrec.to_annot()
        generic_seq = crowsetta.formats.seq.GenericSeq(annots=annot_list)
        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = generic_seq.to_df(basename=True)
E       AttributeError: 'GenericSeq' object has no attribute 'to_df'

test_formats/test_seq/test_generic.py:501: AttributeError
___________________________________________________________________________________________________________ TestGenericSeqClass.test_to_df_onset_offset_sample_only ____________________________________________________________________________________________________________

self = <tests.test_formats.test_seq.test_generic.TestGenericSeqClass object at 0x7f8049e0a6b0>
kaggle_phn_paths = [PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/timit_kaggle/dr1-fvmh0/...'/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/timit_kaggle/dr1-fvmh0/si836.phn')]
timit_phn_as_generic_seq_csv = PosixPath('/Users/tessa/Downloads/crowsetta_review/crowsetta/tests/fixtures/../data_for_tests/csv/timit-dr1-fvmh0-phn.csv')

    def test_to_df_onset_offset_sample_only(self,
                                            kaggle_phn_paths,
                                            timit_phn_as_generic_seq_csv):
        """test whether `annot2df` works when
        the annotations have onsets and offsets specified in seconds only.
        To test this we use the 'timit' format."""
        annot_list = [crowsetta.formats.seq.Timit.from_file(phn_path).to_annot()
                      for phn_path in kaggle_phn_paths]
        generic_seq = crowsetta.formats.seq.GenericSeq(annots=annot_list)
        # below, set basename to True so we can easily run tests on any system without
        # worrying about where audio files are relative to root of directory tree
>       df_created = generic_seq.to_df(basename=True)
E       AttributeError: 'GenericSeq' object has no attribute 'to_df'

test_formats/test_seq/test_generic.py:521: AttributeError
=========================================================================================================================== short test summary info ============================================================================================================================
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_s_only - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_s_and_ind - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestAnnot2DfFunction::test_annot2df_onset_offset_sample_only - AttributeError: module 'crowsetta.formats.seq.generic' has no attribute 'annot2df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_annot2df_onset_offset_s_only - AttributeError: 'GenericSeq' object has no attribute 'to_df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_to_df_onset_offset_s_and_ind - AttributeError: 'GenericSeq' object has no attribute 'to_df'
FAILED test_formats/test_seq/test_generic.py::TestGenericSeqClass::test_to_df_onset_offset_sample_only - AttributeError: 'GenericSeq' object has no attribute 'to_df'
================================================================================================================= 6 failed, 10483 passed in 135.09s (0:02:15) ==================================================================================================================
```
