### 2023-02-10 vocalpy/crowsetta#218 @NickleDave
Hi @rhine3 thank you for catching this.

Could you say a little more about how you installed `crowsetta`?
Did you run `pip install crowsetta` or `conda install crowsetta -c conda-forge`?

Or did you install the development code directly into the environment, e.g. with `pip install -e .`?

If you installed of a package manager, then I think what might be going on is that you have a previous release, that does not include changes I made on the main branch.

Could you please try setting up a development environment as described [here](https://crowsetta.readthedocs.io/en/latest/development/contributors.html#setting-up-a-development-environment) and then running the tests with `nox`? (`nox -s tests`)
Please let me know if you still get the same error then.

I'm not sure what else could cause it besides installing from a package manager. I can see [the CI is passing on main](https://github.com/vocalpy/crowsetta/actions/runs/4038800577) and I do not get this error when I run tests locally.

>     you may want to update the sample issue text here to say "You can determine this in Python by importing crowsetta and running crowsetta.__version__"

Thank you, this is a good idea, I will add it to the issue template
