### 2023-02-10 58bb6f1
DEV: Fix snippet in bug_report.md to print version [skip ci]
as suggested by @rhine3
