### 2023-02-10 vocalpy/crowsetta#218 @rhine3
Hi @NickleDave!

Long story short, you're right that I hadn't tried installing the development environment. I was just testing `crowsetta` directly from within the tests folder within an Anaconda environment that had `crowsetta` installed.

(I can't remember if the environment I was working with had `crowsetta` installed via `pip` or `conda`; I tried both of them in separate environments to verify they both worked.)

Once I actually got the development environment installed, all tests succeeded, except the command I had to run to activate the environment was `. ./.venv/bin/activate` (the documentation says to use the command `. ./.venv/activate`.

I did get one warning when running the tests:

```
../src/crowsetta/_vendor/textgrid/textgrid.py:562
  /Users/tessa/Downloads/crowsetta_review/crowsetta/src/crowsetta/_vendor/textgrid/textgrid.py:562: DeprecationWarning: invalid escape sequence \w
    m = re.match('File type = "([\w ]+)"', header)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
```
