### 2023-02-10 vocalpy/crowsetta#220 @rhine3
BUG: Some Vignettes don't run
**Describe the bug**
While testing the package for PyOpenSci I had some trouble getting the following vignettes to run. To keep this organized in one compact issue, I'll use the "To Reproduce" section to describe the issues.

**To Reproduce**
* Create `conda` environment: `conda create -n crowsetta_review`
* Activate environment: `conda activate crowsetta_review`
* Use `conda` to install Crowsetta from `conda-forge`: `conda install crowsetta -c conda-forge`
* Open Python interpreter in terminal: `python`
* Copy/paste the vignette text into the Python interpreter

What happened for each vignette:
- Vignette “How to convert sequence-like annotations to the 'simple-seq' format” runs successfully locally.
- Vignette “How to use crowsetta with your own annotation format” didn’t run successfully locally because the example file `bat1_annotation.mat` was not included in the tutorial
- Vignette “How to convert any sequence-like format to 'generic-seq'” can’t be run entirely locally. I get an error when running the following line in two places in the vignette:
```
df = generic.to_df()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: 'GenericSeq' object has no attribute 'to_df'
```
- Vignette “How to use crowsetta to remove unlabeled intervals from TextGrid annotations” can’t be run exactly as written locally because the standard conda install of crowsetta didn't include one of the required packages for installing the data, `osfclient`. Perhaps just add an “install osfclient” suggestion to the vignette and that would be good to go. Also, I found that the large code block containing functions wouldn't correctly copy/paste into the interpreter; not sure why.
```
>>>     # set of (unique) states
>>>     states = sorted(
  File "<stdin>", line 1
    states = sorted(
IndentationError: unexpected indent
```

**Expected behavior**
See above.

**Screenshots**
Not applicable.

 - Operating System: Mac
 - Version: `4.0.0.post2`

**Additional context**
I created this issue while performing the [PyOpenSci review of this package](https://github.com/pyOpenSci/software-submission/issues/68).
