### 2023-02-10 #68 @rhine3
---

##### Function documentation:
A Question about pyOpenSci in general - is an individual example really required for every function? The functions are well-documented but some do not have examples, e.g. [crowsetta.data.data.available_formats()](https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.data.data.available_formats.html). Some functions do have examples e.g. [some of the functions here](https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Sequence.html). I’m not sure how to quickly and reliably identify which functions are user-facing and non-user facing. I do feel that the number of examples provided is sufficient.

###### Community guidelines:
- [ ] Great guidelines! These should probably be linked to in the relevant section of the README as well, unless I missed it?

###### README contents:
- [ ] The four vignettes are all linked to using a single link. I think that is okay. It could also be nice to add some additional details near that link, though! Otherwise, The package description is sufficiently clear and well-documented in the links such that it is not necessary to add any additional demonstration of package use in the README.
- [ ] A couple of badges are missing (Repostatus and Python versions supported)

###### Usability:
Overall, the package was very easy to install on my computer. Documentation is linked to on the repository page and in the README. The entire API is fully documented and I felt that the number of examples provided was sufficient.

A minor comment: installing the developer version was slightly more difficult–because the package documentation required me to use pipx/nox, I had to update Homebrew and chase down the source of some errors, which ended up taking me about an hour of detective work (and probably another hour of my computer downloading/upgrading packages in the background). For the average user this doesn’t apply. For the developer, this is a minor and routine obstacle.

###### Functionality:
- [ ] Functionality: Some vignettes didn't run locally. I made an issue for this (https://github.com/vocalpy/crowsetta/issues/220)

###### Tests:
A handful of tests failed at first when I attempted to use a `conda`-installed older version of `crowsetta`. All tests eventually succeeded when I eventually used a development environment (https://github.com/vocalpy/crowsetta/issues/218). I wonder if PyOpenSci’s review template could suggest that reviewers make sure to look in the package's documentation to figure out how the package maintainers request tests be run, e.g., using a development install of the package.

###### Code format:
- [ ] I am unsure how to check/filter actions to see whether there is CI for linting; maybe @NickleDave could direct me to this?

###### JOSS paper
- [ ] There is no `paper.md` in the repository.
