### 2023-02-16 vocalpy/crowsetta#221 @shaupert
BUG: test in crowsetta/tests/test_data.py failed
#### Before submitting a bug, please make sure the issue hasn't been already addressed by searching through [the past issues](https://github.com/NickleDave/vak/issues)

**Describe the bug**
While running tests with Pytest I found that 1 of the tests in crowsetta/tests/test_data.py failed on my colab session.
It looks like the problem comes from : `fixture 'tmp_path' not found`

**To Reproduce**
I cloned the repo on my colab session and run the test with the following command `pytest crowsetta/tests/test_data.py`

**Screenshots**
Errors are reproduced in text below

```
rootdir: /content, inifile:
plugins: typeguard-2.7.1
collected 43 items

crowsetta/tests/test_data.py .E......................................... [100%]

==================================== ERRORS ====================================
___________ ERROR at setup of test_extract_data_files_user_specified ___________
file /content/crowsetta/tests/test_data.py, line 51
  def test_extract_data_files_user_specified(tmp_path):
E       fixture 'tmp_path' not found
>       available fixtures: a_bboxes_list, a_das_csv_path, a_kaggle_phn_path, a_kaggle_wrd_path, a_nist_phn_path, a_nist_wrd_path, a_notmat_path, a_phn_path, a_raven_txt_file, a_seq, a_simple_csv_path, a_textgrid_path, a_transcript_path, a_wrd_path, add_crowsetta, an_audtxt_path, audtxt_paths, birdsong_rec_root, birdsong_rec_wav_path, birdsong_rec_xml_file, birdsongrec_as_generic_seq_csv, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, csv_missing_fields_in_header, csv_root, csv_with_invalid_fields_in_header, csv_with_no_onset_or_offset_column, csv_with_onset_s_but_no_offset_s, csv_with_onset_sample_but_no_offset_sample, das_csv_paths, default_user_data_dir, different_list_of_segments, different_seq, doctest_namespace, example_user_format_annotation_file, example_user_format_as_generic_seq_csv, example_user_format_root, example_user_format_script, kaggle_phn_paths, list_of_segments, monkeypatch, notmat_as_generic_seq_csv, notmat_paths, notmats_root, pytestconfig, raven_dataset_annot_col, raven_root, raven_txt_file_with_no_rows, raven_txt_files, record_property, record_xml_attribute, record_xml_property, recwarn, same_seq, simple_csv_paths, test_data_root, textgrid_paths, timit_kaggle_root, timit_nist_root, timit_phn_as_generic_seq_csv, tmpdir, tmpdir_factory, yarden_annot_mat
>       use 'pytest --fixtures [testpath]' for help on them.
```

/content/crowsetta/tests/test_data.py:51
====================== 42 passed, 1 error in 1.14 seconds ======================

 - Operating System: Linux
 - Version : 4.0.0.post2

**Additional context**
I created this issue while performing the https://github.com/pyOpenSci/software-submission/issues/68.
