### 2023-02-16 #68 @shaupert
Hi, please find my review below:

Estimated hours spent reviewing: 10h

---

This is my first review of a Python package, this is the reason why it took me about 10h.
I would like to say that I learnt a lot during this process, mostly because the `crowsetta` package is really well written (@NickleDave did a really good job) following good practices that, I must confess, I was not aware of. I will try to apply these good practices for my past and next packages as I think it really improves the quality and robustness of a package.

The package does a simple task, which is to parse and convert annotations done on a spectrogram or a waveform into an object that is easy to manipulate. I’m thinking of using it in my own package. As a contributor to the packages [`scikit-maad`](https://github.com/scikit-maad/scikit-maad) and [`bambird`](https://github.com/ear-team/bambird), we had to face different formats of annotation, and `crowsetta` can do the job.

As a user of Audacity to annotate my spectrograms, I was a little bit surprised that `crowsetta` only supports the standard format of Audacity along the time axis. As there is the possibility to register your own format, I tried to figure out how to do that by simply adapting the Raven class and it seems to work. I will add the code of the function later a

Finally I tried to follow the guidelines to create and a class for a new format. I found the process a bit hard to follow, especially because it was written for seq not for bbox which I realized to late. Maybe, it would be interesting to have a guideline for both types of format.

Below are some minor points

##### Test
Only one test failed. I opened an [issue](https://github.com/vocalpy/crowsetta/issues/221)

##### Bugs
crowsetta.data.data.available_formats() and crowsetta.formats.as_list() do not give the same list. The format "yarden" is not in crowsetta.data.data.available_formats()

##### Documentation
crowsetta.Segment : miss a description of the input arguments

##### Improvements
Examples are not provided for each class. I recommend adding a simple example for each class and for each method if possible.
The function scribe cannot load an audacity label .txt file that contains also frequency min and max. I recommend adding a new class to be able to handle this non standard format (see my [issue](https://github.com/vocalpy/crowsetta/issues/222))

##### Summary

I agree with most of the comments of @rhine3. I don't think it's necessary to add them twice ;).
`crowsetta` package deserve the pyOpenScience badge :)
