### 2023-02-16 #68 @NickleDave
Hi @rhine3 and @shaupert thank you both for your detailed reviews and the helpful issues you raised.

I plan to reply here and address those issues within two weeks [as required by the guide](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#response-to-reviews).

Just wanted to let you and @cmarmo know in the meantime that I have seen this 🙏 and I'm working on it 😅
