### 2023-02-27 vocalpy/crowsetta#220 @NickleDave
Hi @rhine3 sorry for the slow reply on this.

Could you try running the vignettes with the development install?

The source of the issue is the same as in #218 -- the version of `crowsetta` that is available on package indexes is behind the main branch.

My fault for not being clearer about that. I think you are right that pyOpenSci should give reviewers more guidance about which versions to install but we can discuss somewhere not on this issue :stuck_out_tongue:
