### 2023-02-27 vocalpy/crowsetta#221 @NickleDave
Hi @shaupert could you please try running tests with the code in the main branch, which is ahead of the releases on PyPI and conda? Using a development environment set up as described here: https://crowsetta.readthedocs.io/en/latest/development/contributors.html#setting-up-a-development-environment

I think the source of the issue is the same as @rhine3 had in #218 and #220. Because I added vignettes + related features, but did not release a new version since I knew I'd be making even more changes during the review! Partly my fault for not making that clearer. And partly because pyOpenSci guides don't have explicit language about this right now as @rhine3 points out
