### 2023-02-27 vocalpy/crowsetta#222 @NickleDave
Thank you very much @shaupert! I really appreciate your help, will start with your implementation and make sure to give you credit.

I'm going to close this as a duplicate of #213 but like I said your code above is a great start and I will be sure to reference this issue there and add you as a contributor.
