### 2023-02-27 vocalpy/crowsetta#218 @NickleDave
Fixed the snippet that shows how to print `crowsetta.__version__` in 58bb6f1
