### 2023-02-27 aca85f1
CLN: Make two regex strings raw to suppress warning
