+++ src/crowsetta/_vendor/textgrid/textgrid.py
@@ -559,7 +559,7 @@ def parse_line(line, short, to_round):
 
 def parse_header(source):
     header = source.readline()  # header junk
-    m = re.match('File type = "([\w ]+)"', header)
+    m = re.match(r'File type = "([\w ]+)"', header)
     if m is None or not m.groups()[0].startswith('ooTextFile'):
         raise TextGridError('The file could not be parsed as a Praat text file as it is lacking a proper header.')
 
@@ -807,7 +807,7 @@ class MLF(object):
 
         source.readline()  # header
         while True:  # loop over text
-            name = re.match('\"(.*)\"', source.readline().rstrip())
+            name = re.match(r'\"(.*)\"', source.readline().rstrip())
             if name:
                 name = name.groups()[0]
                 grid = TextGrid(name)
