### 2023-02-27 vocalpy/crowsetta#218 @NickleDave
Made strings for regex raw to fix DeprecationWarning, in aca85f1
