### 2023-02-27 vocalpy/crowsetta#218 @NickleDave
@all-contributors please add @rhine3 for doc and bug
