### 2023-02-27 b5e94ed
DOC: Fix all contributors badge in README [skip ci]
