### 2023-02-27 075f3e3
DOC: Add Python versions supported badge to README [skip ci]
