### 2023-02-27 8f7f96b
DEV: Add .pre-commit-config.yaml
