+++ .pre-commit-config.yaml
@@ -1,0 +1,34 @@
+repos:
+
+# Autoremoves unused imports
+- repo: https://github.com/hadialqattan/pycln
+  rev: v1.1.0
+  hooks:
+  - id: pycln
+    args: [--config=pyproject.toml]
+    files: ^src/|^tests/
+
+- repo: https://github.com/pycqa/isort
+  rev: 5.12.0
+  hooks:
+  - id: isort
+    name: isort (python)
+    files: ^src/|^tests/
+
+- repo: https://github.com/psf/black
+  rev: 23.1.0
+  hooks:
+  - id: black
+    args:
+      - --line-length=120
+    files: ^src/|^tests/
+
+- repo: https://github.com/PyCQA/flake8
+  rev: 4.0.1
+  hooks:
+  - id: flake8
+    args:
+      - "--max-line-length=120"
+    files: ^src/|^tests/
+    additional_dependencies: &flake8-dependencies
+    - flake8-bugbear
