### 2023-02-27 5caa7f3
DEV: Add 'lint' session to noxfile, fix #135
