### 2023-02-27 #68 @NickleDave
> I wonder if PyOpenSci’s review template could suggest that reviewers make sure to look in the package's documentation to figure out how the package maintainers request tests be run, e.g., using a development install of the package.

@rhine3 I agree with you. Was just discussing this with @lwasser in the pyOS Slack (putting on my pyOS maintainer hat for a second).

Could you please raise an issue on the [software-peer-review](https://github.com/pyOpenSci/software-peer-review/issues) repo suggesting we add this? I am guessing we will do so, and want to make sure you get credit
