+++ src/crowsetta/data/__init__.py
@@ -1,17 +1,3 @@
 """Module that handles example data built into ``crowsetta``"""
-from . import (
-    birdsongrec,
-    generic,
-    notmat,
-    raven,
-    simple,
-    textgrid,
-    timit,
-)
-
-from .data import (
-    available_formats,
-    ExampleAnnotFile,
-    extract_data_files,
-    get,
-)
+from . import birdsongrec, generic, notmat, raven, simple, textgrid, timit
+from .data import ExampleAnnotFile, available_formats, extract_data_files, get
