+++ pyproject.toml
@@ -38,7 +38,8 @@ classifiers = [
 [project.optional-dependencies]
 test = [
     "pytest >=6.2.1",
-    "pytest-cov >=2.12.0"
+    "pytest-cov >=2.12.0",
+    "pytest-xdist >=3.2.0",
 ]
 doc = [
     "ipython != 8.7.0",
