### 2023-02-27 c7b4ecb
DEV: Make dev dependencies include those for doc + test
