### 2023-02-27 e9a1c5a
CLN: Make changes suggested by pycln
