+++ src/crowsetta/interface/__init__.py
@@ -4,3 +4,9 @@ loaded from a file in a specific format."""
 from .base import BaseFormat
 from .bbox import BBoxLike
 from .seq import SeqLike
+
+__all__ = [
+    'BaseFormat',
+    'BBoxLike',
+    'SeqLike'
+]
