### 2023-02-27 78e1582
CLN: Migrate src/crowsetta/ to black formatting
