+++ src/crowsetta/formats/seq/textgrid.py
@@ -55,18 +55,20 @@ class TextGrid:
     (that can be specified when calling ``TextGrid.to_seq``
     or ``TextGrid.to_annot``).
     """
-    name: ClassVar[str] = 'textgrid'
-    ext: ClassVar[str] = '.TextGrid'
+
+    name: ClassVar[str] = "textgrid"
+    ext: ClassVar[str] = ".TextGrid"
 
     textgrid: textgrid.TextGrid
     annot_path: pathlib.Path
     audio_path: Optional[pathlib.Path] = attr.field(default=None, converter=attr.converters.optional(pathlib.Path))
 
     @classmethod
-    def from_file(cls,
-                  annot_path: PathLike,
-                  audio_path: Optional[PathLike] = None,
-                  ) -> 'Self':
+    def from_file(
+        cls,
+        annot_path: PathLike,
+        audio_path: Optional[PathLike] = None,
+    ) -> "Self":
         """load annotations from .TextGrid file,
         like those generated by the Praat application.
 
@@ -88,14 +90,9 @@ class TextGrid:
 
         tg = textgrid.TextGrid.fromFile(annot_path)
 
-        return cls(textgrid=tg,
-                   annot_path=annot_path,
-                   audio_path=audio_path)
+        return cls(textgrid=tg, annot_path=annot_path, audio_path=audio_path)
 
-    def to_seq(self,
-               interval_tier: int = 0,
-               round_times: bool = True,
-               decimals: int = 3) -> crowsetta.Sequence:
+    def to_seq(self, interval_tier: int = 0, round_times: bool = True, decimals: int = 3) -> crowsetta.Sequence:
         """Convert an IntervalTier from this .TextGrid annotation
         into a ``crowsetta.Sequence``.
 
@@ -136,33 +133,23 @@ class TextGrid:
         """
         intv_tier = self.textgrid[interval_tier]
         if not isinstance(intv_tier, textgrid.IntervalTier):
-            raise ValueError(f'Index specified for IntervalTier was {interval_tier}, '
-                             f'but type at that index was {type(intv_tier)}, not an IntervalTier')
+            raise ValueError(
+                f"Index specified for IntervalTier was {interval_tier}, "
+                f"but type at that index was {type(intv_tier)}, not an IntervalTier"
+            )
 
-        onsets_s = np.asarray(
-            [interval.minTime for interval in intv_tier]
-        )
-        offsets_s = np.asarray(
-            [interval.maxTime for interval in intv_tier]
-        )
-        labels = np.asarray(
-            [interval.mark for interval in intv_tier]
-        )
+        onsets_s = np.asarray([interval.minTime for interval in intv_tier])
+        offsets_s = np.asarray([interval.maxTime for interval in intv_tier])
+        labels = np.asarray([interval.mark for interval in intv_tier])
 
         if round_times:
             onsets_s = np.around(onsets_s, decimals=decimals)
             offsets_s = np.around(offsets_s, decimals=decimals)
 
-        seq = crowsetta.Sequence.from_keyword(labels=labels,
-                                              onsets_s=onsets_s,
-                                              offsets_s=offsets_s)
+        seq = crowsetta.Sequence.from_keyword(labels=labels, onsets_s=onsets_s, offsets_s=offsets_s)
         return seq
 
-    def to_annot(self,
-                 interval_tier: int = 0,
-                 round_times: bool = True,
-                 decimals: int = 3
-                 ) -> crowsetta.Annotation:
+    def to_annot(self, interval_tier: int = 0, round_times: bool = True, decimals: int = 3) -> crowsetta.Annotation:
         """Convert an IntervalTier from this .TextGrid annotation
         to a ``crowsetta.Annotation``.
 
@@ -201,10 +188,6 @@ class TextGrid:
[…]
