+++ src/crowsetta/formats/seq/timit.py
@@ -19,6 +19,7 @@ class TimitTranscriptSchema(pandera.SchemaModel):
     """A ``pandera.SchemaModel`` that validates ``pandas`` dataframes
     loaded from a .phn or .wrd file in the TIMIT transcription format.
     """
+
     begin_sample: Optional[Series[int]] = pandera.Field()
     end_sample: Optional[Series[int]] = pandera.Field()
     text: Series[pd.StringDtype] = pandera.Field(coerce=True)
@@ -56,20 +57,18 @@ class Timit:
     audio_path : str. pathlib.Path
         Path to audio file that the TIMIT transcription file annotates.
     """
-    name: ClassVar[str] = 'timit'
-    ext: ClassVar[str] = ('.phn', '.PHN', '.wrd', '.WRD')
+
+    name: ClassVar[str] = "timit"
+    ext: ClassVar[str] = (".phn", ".PHN", ".wrd", ".WRD")
 
     begin_samples: np.ndarray = attr.field(eq=attr.cmp_using(eq=np.array_equal))
     end_samples: np.ndarray = attr.field(eq=attr.cmp_using(eq=np.array_equal))
     text: np.ndarray = attr.field(eq=attr.cmp_using(eq=np.array_equal))
     annot_path: pathlib.Path
-    audio_path: Optional[pathlib.Path] = attr.field(default=None,
-                                                    converter=attr.converters.optional(pathlib.Path))
+    audio_path: Optional[pathlib.Path] = attr.field(default=None, converter=attr.converters.optional(pathlib.Path))
 
     @classmethod
-    def from_file(cls,
-                  annot_path: PathLike,
-                  audio_path: Optional[PathLike] = None) -> 'Self':
+    def from_file(cls, annot_path: PathLike, audio_path: Optional[PathLike] = None) -> "Self":
         """Load annotations from a TIMIT transcription file
 
         Parameters
@@ -99,31 +98,30 @@ class Timit:
         crowsetta.validation.validate_ext(annot_path, extension=cls.ext)
 
         #  assume file is space-separated with no header
-        df = pd.read_csv(annot_path,  sep=' ', header=None)
-        df.columns = ['begin_sample', 'end_sample', 'text']
+        df = pd.read_csv(annot_path, sep=" ", header=None)
+        df.columns = ["begin_sample", "end_sample", "text"]
         df = TimitTranscriptSchema.validate(df)
 
         if audio_path is None:
-            for ext in ('.wav', '.WAV'):
+            for ext in (".wav", ".WAV"):
                 audio_path = annot_path.parent / (annot_path.stem + ext)
                 if audio_path.exists():
                     break
             if not audio_path.exists():
                 # just default to lower-case .wav
-                audio_path = annot_path.parent / (annot_path.stem + '.wav')
+                audio_path = annot_path.parent / (annot_path.stem + ".wav")
 
         return cls(
             annot_path=annot_path,
-            begin_samples=df['begin_sample'].values,
-            end_samples=df['end_sample'].values,
-            text=df['text'].values,
+            begin_samples=df["begin_sample"].values,
+            end_samples=df["end_sample"].values,
+            text=df["text"].values,
             audio_path=audio_path,
         )
 
-    def to_seq(self,
-               round_times: bool = True,
-               decimals: int = 3,
-               samplerate: Optional[int] = None) -> crowsetta.Sequence:
+    def to_seq(
+        self, round_times: bool = True, decimals: int = 3, samplerate: Optional[int] = None
+    ) -> crowsetta.Sequence:
         """Convert this TIMIT annotation to a ``crowsetta.Sequence``.
 
         Parameters
@@ -172,11 +170,11 @@ class Timit:
[…]
