+++ tests/test_formats/test_bbox/test_raven.py
@@ -1,4 +1,3 @@
-import filecmp
 
 import pytest
 
