+++ tests/test_formats/test_seq/test_timit.py
@@ -1,9 +1,8 @@
 import filecmp
 
-import pytest
-
 import pandas as pd
 import pandera
+import pytest
 
 import crowsetta.formats
 
