+++ tests/test_formats/test_seq/test_audtxt.py
@@ -1,10 +1,9 @@
 import filecmp
 
 import numpy as np
-import pytest
-
 import pandas as pd
 import pandera
+import pytest
 
 import crowsetta.formats
 
