+++ tests/fixtures/bbox.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 import pytest
 
 import crowsetta
