+++ tests/test_formats/test_bbox/test_raven.py
@@ -1,8 +1,7 @@
 
-import pytest
-
 import pandas as pd
 import pandera
+import pytest
 
 import crowsetta.formats
 
