### 2023-02-27 3d30e7b
CLN/TST: Apply black formatting to tests/
