+++ tests/test_formats/test_seq/test_timit.py
@@ -8,19 +8,17 @@ import crowsetta.formats
 
 
 class TestTimitTranscriptSchema:
-    def test_timit_schema_phn_df(self,
-                                 a_transcript_path):
-        phn_df = pd.read_csv(a_transcript_path, sep=' ', header=None)
-        phn_df.columns = ['begin_sample', 'end_sample', 'text']
+    def test_timit_schema_phn_df(self, a_transcript_path):
+        phn_df = pd.read_csv(a_transcript_path, sep=" ", header=None)
+        phn_df.columns = ["begin_sample", "end_sample", "text"]
         phn_df = crowsetta.formats.seq.timit.TimitTranscriptSchema.validate(phn_df)
         # if validation worked, we get back a DataFrame
         assert isinstance(phn_df, pd.DataFrame)
 
-    def test_timit_schema_bad_df(self,
-                                 a_transcript_path):
-        phn_df = pd.read_csv(a_transcript_path, sep=' ', header=None)
+    def test_timit_schema_bad_df(self, a_transcript_path):
+        phn_df = pd.read_csv(a_transcript_path, sep=" ", header=None)
         # notice: wrong column names
-        phn_df.columns = ['text',  'begin_sample', 'end_sample']
+        phn_df.columns = ["text", "begin_sample", "end_sample"]
         with pytest.raises(pandera.errors.SchemaError):
             phn_df = crowsetta.formats.seq.timit.TimitTranscriptSchema.validate(phn_df)
 
@@ -48,8 +46,7 @@ def test_to_annot(a_transcript_path):
     assert isinstance(annot, crowsetta.Annotation)
 
 
-def test_to_timit(a_transcript_path,
-                  tmp_path):
+def test_to_timit(a_transcript_path, tmp_path):
     phn = crowsetta.formats.seq.Timit.from_file(annot_path=a_transcript_path)
     annot_path = tmp_path / a_transcript_path.name
     phn.to_file(annot_path=annot_path)
