+++ tests/fixtures/timit.py
@@ -3,7 +3,7 @@ import pytest
 
 from .data import TEST_DATA_ROOT
 
-TIMIT_KAGGLE_ROOT = TEST_DATA_ROOT / 'timit_kaggle'
+TIMIT_KAGGLE_ROOT = TEST_DATA_ROOT / "timit_kaggle"
 
 
 @pytest.fixture
@@ -11,7 +11,7 @@ def timit_kaggle_root(test_data_root):
     return TIMIT_KAGGLE_ROOT
 
 
-KAGGLE_PHN_PATHS = sorted(TIMIT_KAGGLE_ROOT.glob('**/*.phn'))
+KAGGLE_PHN_PATHS = sorted(TIMIT_KAGGLE_ROOT.glob("**/*.phn"))
 
 
 @pytest.fixture
@@ -24,7 +24,7 @@ def a_kaggle_phn_path(request):
     return request.param
 
 
-KAGGLE_WRD_PATHS = sorted(TIMIT_KAGGLE_ROOT.glob('**/*.wrd'))
+KAGGLE_WRD_PATHS = sorted(TIMIT_KAGGLE_ROOT.glob("**/*.wrd"))
 
 
 @pytest.fixture(params=KAGGLE_WRD_PATHS)
@@ -32,7 +32,7 @@ def a_kaggle_wrd_path(request):
     return request.param
 
 
-TIMIT_NIST_ROOT = TEST_DATA_ROOT / 'timit_nist'
+TIMIT_NIST_ROOT = TEST_DATA_ROOT / "timit_nist"
 
 
 @pytest.fixture
@@ -40,7 +40,7 @@ def timit_nist_root():
     return TIMIT_NIST_ROOT
 
 
-NIST_PHN_PATHS = sorted(TIMIT_NIST_ROOT.glob('**/*.PHN'))
+NIST_PHN_PATHS = sorted(TIMIT_NIST_ROOT.glob("**/*.PHN"))
 
 
 @pytest.fixture(params=NIST_PHN_PATHS)
@@ -48,7 +48,7 @@ def a_nist_phn_path(request):
     return request.param
 
 
-NIST_WRD_PATHS = sorted(TIMIT_NIST_ROOT.glob('**/*.WRD'))
+NIST_WRD_PATHS = sorted(TIMIT_NIST_ROOT.glob("**/*.WRD"))
 
 
 @pytest.fixture(params=NIST_WRD_PATHS)
