+++ tests/fixtures/raven.py
@@ -2,7 +2,7 @@ import pytest
 
 from .data import TEST_DATA_ROOT
 
-RAVEN_ROOT = TEST_DATA_ROOT / 'raven'
+RAVEN_ROOT = TEST_DATA_ROOT / "raven"
 
 
 @pytest.fixture
@@ -11,13 +11,12 @@ def raven_root():
     return RAVEN_ROOT
 
 
-ALL_RAVEN_TXT_FILES = sorted(
-    RAVEN_ROOT.joinpath('chronister-at-al-2021/Annotation_Files/Recording_1').glob('*.txt')
-)
+ALL_RAVEN_TXT_FILES = sorted(RAVEN_ROOT.joinpath("chronister-at-al-2021/Annotation_Files/Recording_1").glob("*.txt"))
 # skip first file because it has no annotated rows
 # (but keep it to use for testing error handling)
 RAVEN_TXT_FILE_WITH_NO_ROWS = ALL_RAVEN_TXT_FILES[0]
 RAVEN_TXT_FILES = ALL_RAVEN_TXT_FILES[1:]
+
 
 @pytest.fixture
 def raven_txt_files():
@@ -32,7 +31,7 @@ def a_raven_txt_file(request):
 
 @pytest.fixture
 def raven_dataset_annot_col():
-    return 'Species'
+    return "Species"
 
 
 @pytest.fixture
