+++ pyproject.toml
@@ -59,9 +59,12 @@ doc = [
     "sphinx-tabs >= 3.3.1",
 ]
 dev = [
-    'flit',
+    'black >=23.1.0',
     'crowsetta[doc, test]',
+    'flake8 >=6.0.0',
+    'flit',
     'isort >=5.12.0',
+    'pycln >=2.1.3',
     'twine',
 ]
 
