### 2023-02-27 990afb2
CLN: Make fixes that address flake8 warnings
