+++ src/crowsetta/interface/__init__.py
@@ -1,5 +1,5 @@
 """This module declares the interface for any class that
-represents annotations for animal communication sounds 
+represents annotations for animal communication sounds
 loaded from a file in a specific format."""
 from .base import BaseFormat
 from .bbox import BBoxLike
