+++ src/crowsetta/validation.py
@@ -44,7 +44,7 @@ def check_consistent_length(arrays):
     uniques = np.unique(lengths)
     if len(uniques) > 1:
         raise ValueError(
-            "Found input variables with inconsistent numbers of" " samples: %r" % [int(l) for l in lengths]
+            "Found input variables with inconsistent numbers of" " samples: %r" % [int(length) for length in lengths]
         )
 
 
