+++ src/crowsetta/formats/seq/textgrid.py
@@ -68,7 +68,7 @@ class TextGrid:
         cls,
         annot_path: PathLike,
         audio_path: Optional[PathLike] = None,
-    ) -> "Self":
+    ) -> "Self":  # noqa: F821
         """load annotations from .TextGrid file,
         like those generated by the Praat application.
 
