+++ src/crowsetta/interface/seq/__init__.py
@@ -1,1 +1,5 @@
 from .base import SeqLike
+
+__all__ = [
+    'SeqLike',
+]
