### 2023-02-28 b122b2a
CI: Add GitHub Actions config to run lint on CI
