+++ .github/workflows/lint.yml
@@ -1,0 +1,33 @@
+name: lint
+
+on: [push, pull_request]
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: psf/black@stable
+        with:
+          options: "--line-length=120"
+          src: "./src"
+      - uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+      - name: install libsndfile1 on ubuntu
+        if: matrix.os == 'ubuntu-latest'
+        run: sudo apt install libsndfile1
+      - name: install
+        run: |
+          pip install nox
+          pip install '.[dev]'
+      - name: isort lint
+        uses: isort/isort-action@master
+        with:
+          configuration: "./src"
+      - name: flake8 lint
+        uses: py-actions/flake8@v2
+        with:
+          max-line-length: "120"
+          path: "./src"
+          exclude: "./src/crowsetta/_vendor"
