### 2023-02-28 7b0c963
ENH: Add Audacity extended label track format, fix #213
- Add src/crowsetta/formats/bbox/audbbox.py
- TST: Add tests/data_for_tests/aud-bbox/ with files
- TST: Add audbbox fixtures in tests
- TST: Add tests/test_formats/test_bbox/test_audbbox.py
- Add src/crowsetta/data/audbbox/ with example data
