+++ src/crowsetta/formats/bbox/__init__.py
@@ -1,5 +1,7 @@
+from .audbbox import AudBBox
 from .raven import Raven
 
 __all__ = [
+    "AudBBox",
     "Raven",
 ]
