+++ tests/fixtures/__init__.py
@@ -1,3 +1,4 @@
+from .audbbox import *
 from .audtxt import *
 from .bbox import *
 from .birdsongrec import *
