+++ src/crowsetta/formats/bbox/audbbox.py
@@ -1,0 +1,249 @@
+"""Module for Audacity label tracks
+in extended format, exported to .txt files
+https://manual.audacityteam.org/man/importing_and_exporting_labels.html#Extended_format_with_frequency_ranges
+"""
+from __future__ import annotations
+
+import pathlib
+from typing import ClassVar, List, Optional
+
+import attr
+import pandas as pd
+import pandera
+from pandera.typing import Series
+
+import crowsetta
+from crowsetta.typing import PathLike
+
+
+def txt_to_records(aud_txt_path: PathLike) -> list[dict]:
+    """Load a txt file in Audacity extended label track format
+    into records for a `pandas.DataFrame``.
+
+    Returns a ``list`` of ``dict`` that can be made into a
+    ``DataFrame`` by calling ``pandas.DataFrame.from_records``.
+
+    Parameters
+    ----------
+    aud_txt_path : str, pathlib.Path
+
+    Returns
+    -------
+    records : list
+        Of ``dict``, each ``dict`` a row
+        in the ``DataFrame``.
+
+    Notes
+    -----
+    We work with Audacity txt files this way, instead of
+    loading with ``pandas.read_csv`` then munging, so that we can
+    be sure that we can round-trip data without corrupting it.
+    """
+    with pathlib.Path(aud_txt_path).open('r') as fp:
+        lines = fp.read().splitlines()
+    lines = [line.split('\t') for line in lines]
+
+    records = []
+    # next line: iterate over lines in groups of 2
+    for row1, row2 in zip(*[iter(lines)] * 2):
+        record = {
+            "begin_time_s": float(row1[0]),
+            "end_time_s": float(row1[1]),
+            "label": str(row1[2]),
+            "low_freq_hz": float(row2[1]),
+            "high_freq_hz": float(row2[2]),
+        }
+        records.append(record)
+    return records
+
+
+def df_to_lines(df: pd.DataFrame) -> list[str]:
+    """Convert a pandas DataFrame to a list of strings
+    that can be saved as a txt file in Audacity extended
+    label track format.
+
+    This function is (roughly) the inverse of
+    ``crowsetta.formats.bbox.audbbox.txt_to_records``.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        With contents of a .txt file in Audacity extended label track format,
+        after being loaded and parsed by ``crowsetta.formats.bbox.audbbox.audbbox_txt_to_df``
+
+    Returns
+    -------
+    lines : list
+        List of strings that can be saved to a text file
+        by calling ``writelines``.
+
+    Notes
+    -----
+    We work with Audacity txt files this way, instead of
+    munging and then calling ``pandas.DataFrame.to_csv``,
+    so that we can be sure that we can round-trip data
+    without corrupting it.
+    """
+    df = AudBBoxSchema.validate(df)
+
+    lines = []
+    for record in df.itertuples():
+        row1 = f"{float(record.begin_time_s)}\t{float(record.end_time_s)}\t{record.label}\n"
+        row2 = f"\\\t{float(record.low_freq_hz)}\t{float(record.high_freq_hz)}\n"
+        lines.extend((row1, row2))
+
+    return lines
+
+
+class AudBBoxSchema(pandera.SchemaModel):
+    """A ``pandera.SchemaModel`` that validates ``pandas`` dataframes
+    loaded from Audacity label tracks
+    in extended format, exported to .txt files
+    https://manual.audacityteam.org/man/importing_and_exporting_labels.html#Extended_format_with_frequency_ranges
+    """
+
+    begin_time_s: Series[float] = pandera.Field(coerce=True)
[…]
