### 2023-02-28 vocalpy/crowsetta#221 @shaupert
Hi @NickleDave .
So I tried to figure out how to install brew, pipx and nox on my Ubuntu machine.
Everything worked well following all the provided instructions.
Finally, after cloning crowsetta, I tried to run the command :` nox -s dev`
But I got the following error for numba:

```
Collecting numba>=0.51.0
  Using cached numba-0.56.4.tar.gz (2.4 MB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [8 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/tmp/pip-install-j6nwlu35/numba_19281ba753714f67950a0aed649beb71/setup.py", line 51, in <module>
          _guard_py_ver()
        File "/tmp/pip-install-j6nwlu35/numba_19281ba753714f67950a0aed649beb71/setup.py", line 48, in _guard_py_ver
          raise RuntimeError(msg.format(cur_py, min_py, max_py))
      RuntimeError: Cannot install on Python version 3.11.2; only versions >=3.7,<3.11 are supported.
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.

[notice] A new release of pip is available: 23.0 -> 23.0.1
[notice] To update, run: python -m pip install --upgrade pip
nox > Command /home/haupert/Downloads/crowsetta/.venv/bin/python -m pip install -e '.[dev,test,doc]' failed with exit code 1
nox > Session dev failed.
```
After downgrading my version of Python to 3.9.16, I still have the same error and could no go further as I don't know how to use nox.

You have to know that I'm using `conda` and I created a virtual environment dedicated to crowsetta and lauching the different tools from this virtual env.

I also tried again to run the test directly on a colab session after cloning crowsetta (`!git clone https://github.com/vocalpy/crowsetta.git`) and install crowsetta from the clone (`!pip install /content/crowsetta`) and still have the same issue as before with `tmp_path not found`.

I'm afraid, I can not help you more on that issue. I did my best. I don't think the problem is big, I'm sure you will be able to fix it.

Sylvain
