### 2023-02-28 vocalpy/crowsetta#221 @NickleDave
Thank you @shaupert for taking the time to test this.

I think what's going on with `nox -s dev` is that by default `nox` uses whichever version of Python that `pipx` installs it with.
Then `nox` tries to install `numba` in a way that won't work on your machine.

It's a bit confusing but under the hood `nox` uses `virtualenv` to run different versions of Python; it's not related to your system version. (I had to figure this out from GitHub issues, it's not well documented.)

I did just push a commit that tells `nox` to use Python 3.10 instead for the `dev` session--it should work now, I think. You're on Mac? I was able to set up on a Mac yesterday.

I don't think the tests will ever work for a package installed off of an index, since the distribution package doesn't include (1) pytest as a dependency or (2) the test data, which would make the size of the distribution much bigger.
The error you're getting is because `pytest` can't find one of its own built in fixtures! I guess you might fix that if you update the `pytest` version but you'd still need the test data.

Like you said, it's not the end of the world if you can't run the tests for the review. They're running on CI and passing so we know they must be working somewhere :slightly_smiling_face: Just telling you since I thought you might be interested in what's going on here, or at least what I think is going on :exploding_head:

We did have some discussion for the PyOS packaging guide about whether people should include tests. Since PyPI serves as a kind of "source" for other package managers (e.g. Debian) then it can be good if packages include tests. But IIUC we would really run up the PyPI bills if we uploaded a ton of giant packages, not to mention increasing our carbon footprint. Anyway, not relevant for the review here. I'll go ahead and close this.
