### 2023-02-28 vocalpy/crowsetta#221 @shaupert
Just to let you know that I'm on Ubuntu 22.04.02.
Thanks for your answer :)
