### 2023-02-28 vocalpy/crowsetta#221 @NickleDave
> Just to let you know that I'm on Ubuntu 22.04.02.

Thanks! No offense meant by suggesting you're a Mac user :smirk_cat:

Looks like `numba` for Python 3.11 on any platform is still a WIP although they are getting close: https://github.com/numba/numba/issues/8304#issuecomment-1437292634
