### 2023-02-28 vocalpy/crowsetta#220 @rhine3
Hi David, of course! Did you add a download to `bat1_annotation.mat` on “How to use crowsetta with your own annotation format”?
