### 2023-02-28 vocalpy/crowsetta#220 @NickleDave
Ah, sorry, I should have read more carefully the "What happened for each vignette" you provided.

I will add a download link to that vignette with the file--doing this now.
