### 2023-02-28 vocalpy/crowsetta#220 @NickleDave
@rhine3 added that link to download now.

Writing a to-do list based on your detailed comment abouve:
- [x] include mat file in howto-user-annot.md
- [x] `to_df` method missing should be fixed by local install of code / new release of `crowsetta` after review
- [ ] add `install osfclient` to "unlabeled intervals" vignette
- [ ] fix copy-paste error with code in "unlabeled intervals" vignette
