### 2023-03-01 b631502
CLN: Make COLUMNS_MAP in Raven module a ClassVar
