+++ src/crowsetta/interface/bbox/__init__.py
@@ -1,5 +1,5 @@
 from .base import BBoxLike
 
 __all__ = [
-    'BBoxLike',
+    "BBoxLike",
 ]
