### 2023-03-01 ddfaaba
DEV: Rewrite nox session lint to match what CI does
