### 2023-03-01 2d57103
DOC: Fix vignettes for PyOpenSci review, fix #220
