### 2023-03-01 89aa3d0
DOC: Update CHANGELOG after merging #226 [skip ci]
