### 2023-03-01 3b9c9e0
DOC: Add examples to docstrings in data and formats
