### 2023-03-01 78e5f4c
DOC: Link to CONTRIBUTING and CODE_OF_CONDUCT in README
