### 2023-03-01 0af4560
DOC: Add examples, fix formatting in Sequence docstring
