### 2023-03-01 #68 @NickleDave
Hi all, replying to reviews.

I will reply in separate comments to hopefully make it a bit easier for each reviewer to read.

Starting with @rhine3's review:

I'm focusing here on any concerns raised (but I appreciate the overall positive review!).

> The functions are well-documented but some do not have examples, e.g. [crowsetta.data.data.available_formats()](https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.data.data.available_formats.html). Some functions do have examples e.g. [some of the functions here](https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Sequence.html).

I added examples for these in https://github.com/vocalpy/crowsetta/pull/230

> Great guidelines! These should probably be linked to in the relevant section of the README as well, unless I missed it?

I added links to these in the README in https://github.com/vocalpy/crowsetta/pull/230

> The four vignettes are all linked to using a single link. I think that is okay. It could also be nice to add some additional details near that link, though!

Can you please specify where to add detail? Do you mean on the main index page on the docs?

> A couple of badges are missing (Repostatus and Python versions supported)

Added these in https://github.com/vocalpy/crowsetta/commit/055414306d9423ff3134575d4828f6c2f9dfff29 and https://github.com/vocalpy/crowsetta/commit/075f3e39364a93dd154a410be9b2cbf1ffb9fe28

> A minor comment: installing the developer version was slightly more difficult–because the package documentation required me to use pipx/nox, I had to update Homebrew and chase down the source of some errors, which ended up taking me about an hour of detective work (and probably another hour of my computer downloading/upgrading packages in the background). For the average user this doesn’t apply. For the developer, this is a minor and routine obstacle.

Thank you for taking the time to do this. There's a trade-off between "not using a Makefile that only runs on Unix" v "using something that's pure Python but has its own bootstrapping issues". Just commenting because I think this is of general interest to all of us as developers + contributors + users.

> Some vignettes didn't run locally. I made an issue for this (https://github.com/vocalpy/crowsetta/issues/220)

Fixed by https://github.com/vocalpy/crowsetta/pull/227 and a couple other commits made directly to main as discussed on that issue

> I am unsure how to check/filter actions to see whether there is CI for linting; maybe @NickleDave could direct me to this?

I added a GitHub Actions workflow for linting in https://github.com/vocalpy/crowsetta/pull/224 and I rewrote the nox session `'lint'` to match what the Actions workflow does in https://github.com/vocalpy/crowsetta/commit/ddfaaba137fbb5406d0b26c0b95a4992eb035e1e. Aside: I tried to use pre-commit but it's very aggressive and not particularly easy to configure. Will finish setting it up at some point.

>
> JOSS paper
>
> There is no paper.md in the repository.

Still need to add this! Will do so in the next couple of days
