### 2023-03-01 #68 @NickleDave
Last but not least, to both @rhine3 and @shaupert:

My apologies for confusing you by not release a patch version that was up-to-date with the main branch before you started reviewing. I assumed you would do all the reviewing from a dev environment, but of course you can't test installs from a package index with the dev code. My mistake.

To avoid something like this happening again in this review I have published a pre-release version with all changes I've made so far during the review: https://pypi.org/project/crowsetta/5.0.0rc1/

fyi pip won't install this one unless you specify it directly: `pip install crowsetta==5.0.0rc1`
(it defaults to the latest stable version, just giving you a heads up if you don't know this already / forgot)
