### 2023-03-02 #68 @shaupert
>  crowsetta.data.data.available_formats() and crowsetta.formats.as_list() do not give the same list. The format "yarden" is not in crowsetta.data.data.available_formats()

It's fine for me, don't add it.

> crowsetta.Segment : miss a description of the input arguments
Examples are not provided for each class. I recommend adding a simple example for each class and for each method if possible.

My concern is the documentation of the segment class, not the sequence class where all the arguments are described. See here https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html or try with 'help'. This is true for the version 4.0.0.post2 and the version 5.0.0rc1

> Examples are not provided for each class. I recommend adding a simple example for each class and for each method if possible.

I'm afraid but when I checked on the documentation website https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html, I did not see any example for the classes annotation, transcriber, segment and bbox. Maybe I missed something.

> The function scribe cannot load an audacity label .txt file that contains also frequency min and max. I recommend adding a new class to be able to handle this non standard format (see my https://github.com/vocalpy/crowsetta/issues/222)

Great, thanks a lot !
