### 2023-03-06 0d12477
DOC: Add more examples to docstrings, fix #234
