### 2023-03-06 150d011
CLN: Remove unused Segment.from_row method, fix #231
