+++ src/crowsetta/segment.py
@@ -1,24 +1,21 @@
 """A class that represents a segment of a time series,
 used to annotate animal communication."""
-import warnings
-
 import attr
+import numpy as np
+from attr.validators import instance_of
 
 
-def float_or_None(val):
-    """converter that casts val to float, or returns None if val is string 'None'"""
-    if val == "None":
-        return None
+def int64_to_int(val):
+    """Converter that converts ``numpy.int64`` to ``int``,
+    returns ``int`` as is, and errors for other values.
+    """
+    if hasattr(val, "dtype"):
+        if val.dtype == np.int64:
+            return int(val)
+    elif isinstance(val, int):
+        return val
     else:
-        return float(val)
-
-
-def int_or_None(val):
-    """converter that casts val to int, or returns None if val is string 'None'"""
-    if val == "None":
-        return None
-    else:
-        return int(val)
+        raise TypeError(f"Invalid type {type(val)} for onset or offset sample: {val}. Must be an integer.")
 
 
 @attr.s(frozen=True)
@@ -32,51 +29,37 @@ class Segment(object):
     _FIELDS = ("label", "onset_s", "offset_s", "onset_sample", "offset_sample")
 
     label = attr.ib(converter=str)
-    onset_s = attr.ib(converter=attr.converters.optional(float_or_None))
-    offset_s = attr.ib(converter=attr.converters.optional(float_or_None))
-    onset_sample = attr.ib(converter=attr.converters.optional(int_or_None))
-    offset_sample = attr.ib(converter=attr.converters.optional(int_or_None))
+    onset_s = attr.ib(validator=attr.validators.optional(instance_of(float)), default=None)
+    offset_s = attr.ib(validator=attr.validators.optional(instance_of(float)), default=None)
+    onset_sample = attr.ib(
+        validator=attr.validators.optional(instance_of(int)),
+        converter=attr.converters.optional(int64_to_int),
+        default=None,
+    )
+    offset_sample = attr.ib(
+        validator=attr.validators.optional(instance_of(int)),
+        converter=attr.converters.optional(int64_to_int),
+        default=None,
+    )
     asdict = attr.asdict
 
-    @classmethod
-    def from_row(cls, row, header=None):
-        if type(row) is list:
-            if header is None:
-                raise ValueError("must provide header when row is a list")
-            row = dict(zip(header, row))
-        elif type(row) is dict:
-            if header is not None:
-                warnings.warn(
-                    "Type of 'row' argument was 'dict' but 'header'"
-                    "argument was not None. Value for header will not "
-                    "be used because keys of dict are used as field from "
-                    "header. To use a different header, convert row to a list: "
-                    ">>> row = list(dict.values())"
-                )
+    def __attrs_post_init__(self):
+        if (self.onset_sample is None and self.offset_sample is None) and (
+            self.onset_s is None and self.offset_s is None
+        ):
+            raise ValueError("must provide either onset_sample and offset_sample, or " "onsets_s and offsets_s")
 
-        row_dict = {}
-        for field in cls._FIELDS:
-            try:
-                row_dict[field] = row[field]
-            except KeyError:
-                raise KeyError(f"missing field {field} in row")
-
-        return cls.from_keyword(**row_dict)
+        if self.onset_sample and self.offset_sample is None:
+            raise ValueError(f"onset_sample specified as {self.onset_sample} but offset_sample is None")
+        if self.onset_sample is None and self.offset_sample:
+            raise ValueError(f"offset_sample specified as {self.offset_sample} but onset_sample is None")
+        if self.onset_s and self.offset_s is None:
+            raise ValueError(f"onset_s specified as {self.onset_s} but offset_s is None")
[…]
