### 2023-03-06 vocalpy/crowsetta#234 @NickleDave
DOC: Add examples in docstrings where missing
- [x] Segment class
- [x] Annotation
- [x] Transcriber
- [x] BBox

As suggested by @shaupert in pyOpenSci review: https://github.com/pyOpenSci/software-submission/issues/68#issuecomment-1451660136

> My concern is the documentation of the segment class, not the sequence class where all the arguments are described. See here https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html or try with 'help'. This is true for the version 4.0.0.post2 and the version 5.0.0rc1
>
>     Examples are not provided for each class. I recommend adding a simple example for each class and for each method if possible.
>
> I'm afraid but when I checked on the documentation website https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html, I did not see any example for the classes annotation, transcriber, segment and bbox. Maybe I missed something.
