### 2023-03-06 d6208a1
CLN: Remove Segment.from_keyword method, not needed
