### 2023-03-06 879ee47
DOC: Fix Examples in crowsetta/bbox.py [skip ci]
