+++ src/crowsetta/bbox.py
@@ -30,8 +30,8 @@ class BBox:
     label : str
         string label that annotates bounding box
 
-    Examples:
-    ---------
+    Examples
+    --------
 
     A toy example of a bounding box-like annotation.
 
@@ -40,9 +40,7 @@ class BBox:
     >>> bboxes = [bbox1, bbox2]
     >>> annot = crowsetta.Annotation(notated_path='prebird1.wav', annot_path='prebird1.csv', bboxes=bboxes)
     >>> print(annot)
-    Annotation(annot_path=PosixPath('prebird1.csv'), notated_path=PosixPath('prebird1.wav'),
-    bboxes=[BBox(onset=1.0, offset=2.0, low_freq=3000.0, high_freq=10000.0, label='Pinacosaurus grangeri'),
-    BBox(onset=3.0, offset=4.0, low_freq=3250.0, high_freq=12500.0, label='Pinacosaurus grangeri')])
+    Annotation(annot_path=PosixPath('prebird1.csv'), notated_path=PosixPath('prebird1.wav'), bboxes=[BBox(onset=1.0, offset=2.0, low_freq=3000.0, high_freq=10000.0, label='Pinacosaurus grangeri'), BBox(onset=3.0, offset=4.0, low_freq=3250.0, high_freq=12500.0, label='Pinacosaurus grangeri')])  # noqa
     """
 
     onset: float = field(validator=is_positive)
