### 2023-03-06 #68 @NickleDave
> My concern is the documentation of the segment class, not the sequence class where all the arguments are described. See here https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html or try with 'help'. This is true for the version 4.0.0.post2 and the version 5.0.0rc1

Apologies, I should have read more closely.

> I'm afraid but when I checked on the documentation website https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Segment.html, I did not see any example for the classes annotation, transcriber, segment and bbox. Maybe I missed something.

I have added examples to the docstrings for all of these classes in https://github.com/vocalpy/crowsetta/pull/235.

Segment: https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Sequence.html
Annotation: https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Annotation.html#crowsetta.Annotation
Bounding Box: https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.BBox.html
Transcriber: https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.Transcriber.html

I also noticed that there was a Segment classmethod that was no longer used, and that I could factor out the other one. So it was a good thing you pushed me to add examples--I would have missed this otherwise. Thank you.
Removed the `Segment.from_row` method here (see linked issue for detail): https://github.com/vocalpy/crowsetta/pull/232
And removed `Segment.from_keyword` in https://github.com/vocalpy/crowsetta/pull/235.
