### 2023-03-06 #68 @NickleDave
@rhine3 @shaupert please do let me know if there are any other issues I need to address.

I do know I need to add the paper.md file for JOSS. Will get to that in the next couple of days.
