### 2023-03-07 6490162
Add JOSS paper.md and paper.bib
