+++ paper.bib
@@ -1,0 +1,556 @@
+
+@article{wirthlin_modular_2019,
+	title = {A {Modular} {Approach} to {Vocal} {Learning}: {Disentangling} the {Diversity} of a {Complex} {Behavioral} {Trait}},
+	volume = {104},
+	issn = {08966273},
+	shorttitle = {A {Modular} {Approach} to {Vocal} {Learning}},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0896627319308396},
+	doi = {10.1016/j.neuron.2019.09.036},
+	language = {en},
+	number = {1},
+	urldate = {2021-05-22},
+	journal = {Neuron},
+	author = {Wirthlin, Morgan and Chang, Edward F. and Knörnschild, Mirjam and Krubitzer, Leah A. and Mello, Claudio V. and Miller, Cory T. and Pfenning, Andreas R. and Vernes, Sonja C. and Tchernichovski, Ofer and Yartsev, Michael M.},
+	month = oct,
+	year = {2019},
+	pages = {87--99},
+	file = {Wirthlin et al. - 2019 - A Modular Approach to Vocal Learning Disentanglin.pdf:/home/pimienta/Zotero/storage/L7JW7EGY/Wirthlin et al. - 2019 - A Modular Approach to Vocal Learning Disentanglin.pdf:application/pdf},
+}
+
+@article{sainburg_toward_2021,
+	title = {Toward a {Computational} {Neuroethology} of {Vocal} {Communication}: {From} {Bioacoustics} to {Neurophysiology}, {Emerging} {Tools} and {Future} {Directions}},
+	volume = {15},
+	issn = {1662-5153},
+	shorttitle = {Toward a {Computational} {Neuroethology} of {Vocal} {Communication}},
+	url = {https://www.frontiersin.org/articles/10.3389/fnbeh.2021.811737/full},
+	doi = {10.3389/fnbeh.2021.811737},
+	abstract = {Recently developed methods in computational neuroethology have enabled increasingly detailed and comprehensive quantiﬁcation of animal movements and behavioral kinematics. Vocal communication behavior is well poised for application of similar large-scale quantiﬁcation methods in the service of physiological and ethological studies. This review describes emerging techniques that can be applied to acoustic and vocal communication signals with the goal of enabling study beyond a small number of model species. We review a range of modern computational methods for bioacoustics, signal processing, and brain-behavior mapping. Along with a discussion of recent advances and techniques, we include challenges and broader goals in establishing a framework for the computational neuroethology of vocal communication.},
+	language = {en},
+	urldate = {2022-01-22},
+	journal = {Frontiers in Behavioral Neuroscience},
+	author = {Sainburg, Tim and Gentner, Timothy Q.},
+	month = dec,
+	year = {2021},
+	pages = {811737},
+	file = {Sainburg and Gentner - 2021 - Toward a Computational Neuroethology of Vocal Comm.pdf:/home/pimienta/Zotero/storage/AZFQEZ85/Sainburg and Gentner - 2021 - Toward a Computational Neuroethology of Vocal Comm.pdf:application/pdf},
+}
+
+@article{kershenbaum_acoustic_2016,
+	title = {Acoustic sequences in non-human animals: a tutorial review and prospectus: {Acoustic} sequences in animals},
+	volume = {91},
+	issn = {14647931},
+	shorttitle = {Acoustic sequences in non-human animals},
+	url = {https://onlinelibrary.wiley.com/doi/10.1111/brv.12160},
+	doi = {10.1111/brv.12160},
[…]
