+++ paper.md
@@ -1,0 +1,168 @@
+---
+title: 'Crowsetta: A Python tool to work with any format for annotating animal vocalizations and bioacoustics data.'
+tags:
+  - animal acoustic communication
+  - bioacoustics
+  - Python
+authors:
+  - name: David Nicholson
+    orcid: 0000-0002-4261-4719
+    corresponding: true
+    affiliation: 1
+affiliations:
+  - name: Independent Research, USA
+    index: 1
+date: 07 Mar 2023
+bibliography: paper.bib
+---
+
+# Summary
+
+Studying how animals communicate with sound allows researchers to answer 
+a wide range of questions, from "What species of birds live in this area?", 
+to "How do mice mothers protect their young?". 
+Some animals learn their vocalizations, and by studying the vocal behavior 
+of these animals we can investigate questions like "How did speech evolve?".
+Answering these questions require computational methods and big team science 
+across many disciplines, 
+such as ecology, ethology, bioacoustics, psychology, neuroscience, linguistics, and genomics.
+
+Analyses of animal acoustic communication often require annotating 
+the sounds animals make. 
+To annotate animal sounds, researchers typically use graphical user interfaces (GUIs),
+that enable them to annotate audio and/or spectrograms.
+Such annotations usually include the times when sound events start and stop, and labels that 
+assign each sound to some set of classes chosen by the annotator.
+GUI applications save the annotations in many different file formats.
+This Python package, crowsetta, can parse the most widely used formats,
+and it provides software abstractions 
+that make it easy to extend the library to parse new formats. 
+In this way, crowsetta allows researchers 
+to work with data annotated in a wider variety of formats. 
+Additionally, crowsetta helps users convert annotations to simple file formats,
+such as csv files, that do not require detailed knowledge of the annotation format itself.
+This facilitates loading the annotations with widely used libraries 
+for data analysis (e.g., Pandas in Python), 
+and also promotes sharing data.
+Overall, crowsetta supports the interdisciplinary collaboration 
+required for the study of animal acoustic communication.
+
+# Statement of need
+
+Studying how animals communicate with sound 
+allows researchers to answer a wide range of questions.
+For example: 
+"What is the language faculty, and how did it evolve?" [@hauser_faculty_2002], 
+and "What is the basis of learned vocalizations in animals?" [@wirthlin_modular_2019]. 
+Answering these questions will require big team science 
+and true interdisciplinary collaborations [@hauser_faculty_2002; @wirthlin_modular_2019].
+The methods used to answer these questions 
+are becoming ever more computational and data driven.
+As one example of such methods, deep learning models
+have recently proliferated in the field of bioacoustics [@stowell_computational_2022] 
+and adjacent fields like animal behavior [@berman_measuring_2018; @pereira_quantifying_2020], 
+specifically studies of vocal behavior [@sainburg_toward_2021]. 
+These models have also become common in the field of neuroscience [@cohen_recent_2022], 
+and more specifically in studies of the neural bases of vocal communication
+[@coffey_deepsqueak_2019; @cohen_automated_2022; @goffinet_low-dimensional_2021; @sainburg_finding_2020; @steinfath_fast_2021].
+The performance of these data-driven deep learning models depends crucially on annotated data.
+High-quality annotations from humans 
+are needed to train supervised learning models to predict annotations
+[@coffey_deepsqueak_2019; @cohen_automated_2022; @steinfath_fast_2021], 
+or to provide inputs to unsupervised models that perform dimensionality reduction 
+and compute measures of similarity [@goffinet_low-dimensional_2021; @sainburg_finding_2020].
+In spite of this clear-cut need for high quality, easily accessible 
[…]
