### 2023-03-09 #68 @cmarmo
@rhine3 @shaupert thank you for all your work so far!
@NickleDave thanks for your follow-up, waiting for the `paper.md` and (hopefully ... 🙂) a final approval from both our reviewers.
In the meanwhile, as we discussed off-line, I have been in touch with @YannickJadoul and he kindly accepted to provide his expertise on TextGrid format and [praat](https://www.fon.hum.uva.nl/praat/) to comment about crowsetta usability in and interoperability with the linguistic/phonetic community. I let him introduce himself here: thanks Yannick for your collaboration! :pray:
