### 2023-03-09 #68 @shaupert
Hi @NickleDave , @rhine3  and @cmarmo,

@NickleDave : thanks for taking into accounts all my observations. You really did a great job.
@cmarmo :  I give my final approval for the package [crowsetta](https://github.com/vocalpy/crowsetta)

I really enjoyed the reviewing process. It took me time as I'm not familiar with the differnet tools. I learnt a lot.

Cheers
