### 2023-03-09 #68 @NickleDave
Hi @cmarmo I just added a branch with the JOSS paper:
https://github.com/vocalpy/crowsetta/tree/add-joss-paper

@shaupert thanks so much for your very helpful review.
