### 2023-03-09 #68 @NickleDave
Hi @YannickJadoul :wave:  would love your advice on the best way to support TextGrid files :pray:.
See issue here: https://github.com/vocalpy/crowsetta/issues/46

I assume in Parselmouth you just wrap the Praat code for loading TextGrids directly. I am of course a huge fan of ParselMouth but adding that as a dependency would be sort of contrary to the goals for crowsetta (since it would bring in all of Praat and secondary dependencies with it, and I'm just trying to handle the file format).

Do you have opinions on any of the (many) pure Python packages that are out there for parsing TextGrid files? I have vendored one and I know of at least one alternative--I haven't personally run major issues with either but I also notice they are a bit rough around the edges.
