### 2023-03-11 #68 @YannickJadoul
Hi all!
Yes, I'll have a look at the crowsetta - Praat interface somewhere in the next couple of days! I'd actually seen this review show up in my mailbox, before, so I'm extra curious to now have the perfect excuse to have a deeper look into crowsetta :-)

> I have vendored one library now to parse the format and I know of at least one alternative that looks fairly comprehensive. See issue here: [vocalpy/crowsetta#46](https://github.com/vocalpy/crowsetta/issues/46).
>
> The main problem is that the code I have vendored now won't load any Praat files where there's overlapping segments. For example, see the dataset I use in this vignette, where I skip over some files because they won't load: https://crowsetta.readthedocs.io/en/latest/howto/remove-silent-labels-textgrid.html But when I tested TextGridTools as an alternative it had any pain points.

[TextGridTools](https://github.com/hbuschme/TextGridTools) is the one I'm most familiar with, but I've never done a full survey. Some of the main points speaking for tgt is that it's very focused on _only_ dealing with this one format, it has a nice and straightforward API, and the developers have been responsive in releasing a new version when I bumped into one minor bug. But again, can't say I've done a full comparative study, and I can't say that the current vendored package doesn't have these or other things going for it!

> I assume in Parselmouth you just wrap the Praat code for loading TextGrids directly. I am of course a huge fan of ParselMouth but adding that as a dependency would be sort of contrary to the goals for crowsetta (since it would bring in all of Praat and secondary dependencies with it, and I'm just trying to handle the file format).

Yes, I do. And actually, TextGrids don't have a full Python API yet, so you have to go through the generic [`parselmouth.praat.call`](https://parselmouth.readthedocs.io/en/stable/api_reference.html#parselmouth.praat.call). The one extra thing I'm providing is `parselmouth.TextGrid.to_tgt` and `parselmouth.TextGrid.from_tgt`, as a way of delegating Pythonic manipulation to TextGridTools :-)
But maybe in the future, there can be a `to_crowsetta`? ;-)

> Do you have opinions on any of the (many) pure Python packages that are out there for parsing TextGrid files? https://pypi.org/search/?q=textgrid (At least ~20 scream by my count, not including crowsetta).

That all being said, zooming out a bit, Praat's file formats tends to be rather stable (especially for things as widespread as TextGrids). So the actual choice whatever Python package you decide on won't matter, I would expect, if it's "just" an internal dependency to crowsetta?
Or, looking at it differently, the main test for all these packages will only be when the Praat developers decide to add something to the TextGrid format. But I also can't predict which of these 20 packages would in such case be the first to update.
