### 2023-03-13 #68 @rhine3
Hi all, on my part, David responded to my review satisfactorily. As for your question David -
> As for can you please specify where to add detail? Do you mean on the main index page on the docs?

I was just thinking it could be good to add a little more detail to the README to make the link to the vignettes "stick out" more prominently. However, looking it over again, I don't think this is necessary.

Thanks so much for all the work you did to address these comments, and thanks for contributing this useful package to the scientific community, @NickleDave! I am happy to give my final approval on the package. Please let me know if anything else is needed!
