### 2023-03-13 #68 @NickleDave
> Thanks so much for all the work you did to address these comments, and thanks for contributing this useful package to the scientific community

Thanks so much @rhine3 for your helpful review, my repo is now up to spec thanks to you (check out those badges 😎 ) and I finally pushed myself to be better about linting too. And your feedback about nox was also quite helpful! Appreciate all the time you put into this.
