### 2023-03-13 #68 @NickleDave
@YannickJadoul thank you for your reply. I will look again at TextGridTools and report back. I take your broader point that the TextGrid format is relatively stable.
