### 2023-03-14 #68 @cmarmo
> Yes, I'll have a look at the crowsetta - Praat interface somewhere in the next couple of days! I'd actually seen this review show up in my mailbox, before, so I'm extra curious to now have the perfect excuse to have a deeper look into crowsetta :-)

@NickleDave, if I understand correctly Yannick was planning some more detailed comments.
