### 2023-03-15 #68 @YannickJadoul
Hi @NickleDave!

I've had a closer look at crowsetta and the TextGrids. And wow, it's indeed a whole lot of work! Curious to try it out on some future project, myself :-)

To first answer your explicit concerns:

> I don't mean to sound like I'm being overly unfair to a package that's clearly working for some chunk of people that are _not_ complaining about it. But spending my "development budget" to switch to it wouldn't help I think. My impression from looking briefly at other textgrid libraries on PyPI is that the same logic would apply there--I wouldn't get a lot out of taking the time to switch, at least not right now. If people using crowsetta start complaining about how it handles TextGrid files then it might be time to look again.

Overall, yes, agreed. All these libraries have the basic functionality, and then a few separate advantages/disadvantages, it seems. And all of them also seem semi-dormant ([textgrid.py](https://github.com/kylebgorman/textgrid)'s last commit also seems to be 3 years ago, but maybe there's just not a lot more to do?). So as long as things "just work" within the context of crowsetta, I don't think I have a strong argument for either of them.
(As said, I just happen to know TextGridTools and its interface better, and I do like their [separate docs and API reference with extra utility functions](https://textgridtools.readthedocs.io/en/stable/api.html) when I have to work with only TextGrids :-) )

> * I found that, just like the code I have vendored now, textgridtools also raises a ValueError when opening most of the files from [this dataset](https://osf.io/r6paq/) because of overlapping tiers. So switching to textgridtools would not help me get around that issue. (My understanding is that overlapping tiers are technically valid although it could be the case there's an issue with the dataset. As described on section 6, "weak invariant" of https://www.fon.hum.uva.nl/praat/manual/TextGrid_file_formats.html)

I've always thought that this is really just violating the TextGrid format. When creating them in Praat, you normally can't have overlapping intervals (which is the reason for these empty intervals, because each tier really just partitioning the whole duration into intervals; when you create an empty tier, it will start already contain one interval spanning everything).
That being said, I'm surprised Praat doesn't fix it when it loads such a broken TextGrid. I've been trying to get Praat to crash or behave badly on such a TextGrid, but haven't managed. All I can say is
1) I don't believe you can make an IntervalTier with overlapping intervals within Praat without changing the saved format.
2) Praat does behave weirdly when this happens, and it's clearly some sort of unforeseen/unsupported situation:
![image](https://user-images.githubusercontent.com/8025744/225439938-6c25595e-9b19-4cb0-80fc-0c4f83767e66.png)
![image](https://user-images.githubusercontent.com/8025744/225439903-21eae378-514f-450b-8262-fdb7a430ad14.png)
Or when converted "To DurationTier...":
![image](https://user-images.githubusercontent.com/8025744/225440758-47d89f67-073a-495e-8fb0-eeed0e3663fe.png)

So I think it would be fine to refuse these things in crowsetta.
Alternatively, I don't really know how to read such a TextGrid without exception. Best I can suggest is to adapt the code you're vendoring inside of crowsetta anyway, and disable this check, here?
https://github.com/kylebgorman/textgrid/blob/19d3ea2f765a435bb22a529602532752361167c9/textgrid/textgrid.py#L350-L351

And then my list of comments and thoughts:
- Maybe [the TextGrid format description](https://crowsetta.readthedocs.io/en/latest/formats/seq/textgrid.html#textgrid) could contain a short paragraph on how the format maps to crowsetta's data model? I.e., one IntervalTier tier gets read into a single `Sequence`-type `Annotation`. Also, the [remove-silent-labels-textgrid](https://crowsetta.readthedocs.io/en/latest/howto/remove-silent-labels-textgrid.html) How-To contains a whole paragraph explaining this mapping (and details such as empty intervals being included), so maybe you either want to link to that, or extract the important parts and have them in the "central TextGrid format description".
- Since you can just read one tier, is there an easy way of merging `Annotations`? E.g., since TextGrids don't work with overlapping intervals, we will often have multiple tiers for multiple animals in the same recording. It's not entirely clear to me if I can merge those all into a single `Annotation`, or whether I should keep multiple `Annotation`  objects.
- In my experience, empty TextGrid intervals are almost always the equivalent of "no annotation" (given the way IntervalTiers work, as mentioned above). So I would suggest that maybe you could remove those by default, or at least provide an easy way for users to not do this manually.
Again, I might be biased here, since [TextGridTools also does it by default](https://textgridtools.readthedocs.io/en/stable/api.html#tgt.io.read_textgrid), but I've found it a perfect default in everyday use (often not even noticing it's happening, since it's so obvious).
- I've tried the various TextGrid formats Praat supports (long, short, binary). No clue why, but e[ven though it would seem to be supported by textgrid.py](https://github.com/kylebgorman/textgrid/tree/19d3ea2f765a435bb22a529602532752361167c9/tests/data), it can't read the unicode-containing short-format TextGrids I fed it (see [TextGrid_formats.zip](https://github.com/pyOpenSci/software-submission/files/10984527/TextGrid_formats.zip)). The issue seems to be that my files (freshly written with Praat 6.2.23) don't have this same starting line: https://github.com/kylebgorman/textgrid/blob/19d3ea2f765a435bb22a529602532752361167c9/tests/data/short_format.TextGrid#L1. I don't know when Praat stopped writing "short" in these files, how TextGridTools handles this, and if this is worth your development time. But it might be good to note in the documentation that not all short TextGrids can be read?
Binary-format TextGrid files cannot be read either, but I don't think this is very important. I don't think I've encountered anything besides Praat (and by transitivity, Parselmouth ;-) ) to read those things. But maybe just specify that only text-based formats are supported.
Encoding-wise, everything seems to be going well, UTF-8 vs. UTF-16 getting autodetected by the vendored library. I'm not a 100% sure if this can give any issues when there's ambiguity in encoding, but I guess it should be fine? (And this is a point for textgrid.py, since TextGridTools needs to get the encoding specified explicitly, if different from UTF-8)
- Mostly out of curiosity: my `Segment.onset_sample` and `Segment.offset_sample` are always `None. Is there a way to specify an audio file to get these filled in when reading a TextGrid? Or is this just not applicable in this case?
- Potentially, when converting a TextGrid to Annotation, it's nice if the user can specify a tier name instead of just the index. But then again, nothing urgent; more a thought of something useful.
- I don't think crowsetta means to do provide this, but I couldn't find a way to write TextGrids, right? So I don't need to have a look at that?

A few more small things I encountered. I didn't do a whole review, but as I saw those, I thought it's worth mentioning:
- The Tutorial page of the docs has an execution error somewhere (`ValueError: Format name 'aud-txt' not recognized.Valid format names:`), which means that most of the that page doesn't have the output listed.
- On the "How to use crowsetta to remove unlabeled intervals from TextGrid annotations" page, there's also something weird happening to the output during the download. Maybe there's a way of suppressing the output of that cell?
- In general, throughout the docs and especially the API reference, things like `crowsetta.Annotation` are not clickable. I think that's a matter of having single ` backticks vs. double `` backticks if you're using Sphinx ([especially if you put a default role of `py:obj`](https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-default_role), I believe). But I noticed how used I am to be able to do that and jump around the API.

So, but overall, I don't see any major issues! Congrats with this beautiful project and all the work that has clearly gone into this! :-)
Happy to think along if there's still anything in particular from your side regarding TextGrids (e.g. going deeper into the overlap issue). But I don't believe this review process deserves to be stalled on just that ;-)
