### 2023-03-17 #68 @NickleDave
Thanks so much @YannickJadoul.

The additional information on the TextGrid file format and TextGridTools is good to know.

And I really appreciate your taking the time to look at that specific dataset; I had a hunch it's only technically valid. Probably I should contact the authors.

Your detailed comments are quite helpful! @cmarmo I expect to have these addressed by the middle of next week but just wanted to acknowledge and thank @YannickJadoul now
