### 2023-03-20 8808e32
DEV/CI: Change nox session 'coverage' to not run tests in parallel
