### 2023-03-20 d9a75b6
BUG: Coerce GenericSeq path attributes to str
Add `coerce=True` to the `annot_path` and `notated_path`
attributes of the GenericSeq pandera schema,
to force them to be converted to strings.

This fixes failures on CI where I guess these
get passed in as Path classes even though when I run
tests locally they are not and tests pass?
Seems like a good idea to force them to be string anyway.
