### 2023-03-20 9c1a5e0
DEV: Fix nox session 'dev' to work on Windows
