### 2023-03-20 b4d1ca4
BUG: Fix how Segment converts on/offset sample to int
The onset/offset sample attributes of Segment
have a converter that was written to convert
numpy int64 to Python int.

But it turns out that we don't always default to int64,
e.g., on Windows it's int32.
The converter failed then since it first checked for a dtype
attribute, then checked *only* for int64.
This logical error caused us to return None for int32.

Instead we check whether there's a dtype *and* it's an instance
of np.integer, which catches any numpy int type.
