### 2023-03-20 b76482b
CLN: Rename 'Hertz' -> 'sample' in Sequence unit tests
