### 2023-03-20 4af3a3c
DOC: Update CHANGELOG after merging #237 and #238 [skip ci]
