### 2023-03-20 da6f385
DOC/CLN: Fix cross-refs in docstrings, style. Fixes #239
Mostly fixing cross-refs in docstrings.
Also makes some minor style/formatting fixes,
e.g., ".txt file" -> "txt file",
adds citations in a couple places.
Add type hints in validation.py
