### 2023-03-20 vocalpy/crowsetta#239 @NickleDave
DOC: Fix/add cross-references in docstrings
so they render as links in the API docs

as suggested in pyOS review here: https://github.com/pyOpenSci/software-submission/issues/68#issuecomment-1470879706
