### 2023-03-21 0fdea17
DOC: Suppress cell output, other clean-ups, in how-to
