### 2023-03-21 vocalpy/crowsetta#241 @NickleDave
ENH: Better handle Praat TextGrids
as discussed here (will update this issue more): https://github.com/pyOpenSci/software-submission/issues/68#issuecomment-1470879706

- [x] Be able to parse both "normal" and short format TextGrid files, in both UTF-8 and UTF-16 encoding
- [x] Remove empty intervals by default, provide option to not remove them
- [x] Be able to load a TextGrid with multiple IntervalTiers and then convert to an Annotation with multiple sequences, this will require fixing #42 first
- [x] Use reprlib with the vendored TextGrid so it is readable
- [x] Do we need to add/update link to spec? As ref'd here: https://github.com/pyOpenSci/software-submission/issues/68#issuecomment-1478474921. I don't think we have this page: https://www.fon.hum.uva.nl/praat/manual/TextGrid_file_formats.html -- I think we have some other one
