### 2023-03-21 #68 @drammock
@NickleDave just a drive-by comment to make sure you're aware of the official TextGrid format spec: https://www.fon.hum.uva.nl/praat/manual/TextGrid_file_formats.html  In particular, this quote:

> ...the union of the time domains of the set of intervals in a tier is a contiguous stretch of time, and no intervals overlap.

So you are free to error out on TextGrids that have overlapping intervals.
