### 2023-03-23 #68 @cmarmo
Hi @NickleDave , just checking... Do you think you will be able to address all the comments about TextGrid, or maybe it is worth to submit to JOSS and let major changes for a new version of crowsetta?
I think it is important that all your work and the review effort come to a well deserved achievement in a reasonable time... :)
Let me know if I can proceed to the editor checks for submission to JOSS, thanks!
