### 2023-03-23 #68 @NickleDave
Hi @cmarmo I think I will be able to address but it's taking a little longer than expected.

Feedback from Yannick has made me realize crowsetta is not quite providing what it promises for TextGrids. I thought the fix would be trivial but it turns out the code I vendored is not so easy to change.

I have a plan to address, just haven't had time yet.
I think I can finish by Monday if that's okay.
