### 2023-03-24 54ee3c9
ENH: Better handle Praat TextGrids, fix #241
* Be able to parse both "normal" and short format TextGrid files,
  in both UTF-8 and UTF-16 encoding
* Remove empty intervals by default,
  provide option to not remove them
* Be able to load a TextGrid with multiple IntervalTiers
  and then convert to an Annotation with multiple sequences
* Use reprlib with the vendored TextGrid so it is readable

Squash commits:
- Remove _vendor/textgrid
- Move textgrid module into sub-package
- Move test_textgrid into its own subpackage
- Change example TextGrid to a file with more than one tier, for demo purpos
[…]
