+++ tests/data_for_tests/textgrid/README.md
@@ -1,0 +1,30 @@
+# sources
+
+## Calhoun et al. 2022
+
+Sound files and Praat TextGrids accompanying Sasha Calhoun, Corinne Seals, Toaga Alefosio and Niusila Faamanatu-Eteuati,
+'Phrasal Prosody of Heritage Speakers of Samoan in Aotearoa New Zealand', 
+to appear in Rajiv Rao (ed), The Phonetics and Phonology of Heritage Languages, Cambridge University Press.
+https://osf.io/bxasd/
+
+These files were used because they have multiple IntervalTiers.
+
+## parselmouth
+
+Files in ./parselmouth were kindly provided by Yannick Jadoul
+during the pyOpenSci review.
+
+They are adaptations of a test TextGrid file used by Parselmouth:
+https://github.com/YannickJadoul/Parselmouth
+Jadoul, Y., Thompson, B., & de Boer, B. (2018). 
+Introducing Parselmouth: A Python interface to Praat. Journal of Phonetics, 71, 1-15. 
+https://doi.org/10.1016/j.wocn.2018.07.001
+
+## praatIO
+To test short format TextGrids we use some text files from PraatIO.
+
+Tim Mahrt. PraatIO. https://github.com/timmahrt/praatIO, 2016.
+
+The code for parsing TextGrid files is itself adapted from this library,
+under MIT license:
+https://github.com/timmahrt/praatIO/blob/main/LICENSE

[…]
