### 2023-03-28 9311d82
DOC/CLN: Fix ref in birdsongrec.py docstring
