### 2023-03-28 44643a0
DOC: Clean/revise src/crowsetta/formats/seq/textgrid/textgrid.py docstring
