### 2023-03-28 vocalpy/crowsetta#244 @NickleDave
DOC: Rewrite 'remove-empty-intervals' tutorial to be an example of an end-to-end analysis with crowsetta
- [ ] Rewrite to just be an example of an end-to-end analysis
- [ ] Add the TextGrid files from the dataset that parse without error to repo, instead of downloading of OSF, to avoid download error and make tutorial shorter
- [ ] Rewrite empty intervals language to explain it's done by default
