### 2023-03-28 #68 @NickleDave
Just merged a final PR that addresses feedback from @YannickJadoul along with some stray commits I made to main that I'll link here:
https://github.com/vocalpy/crowsetta/pull/243

At a high level, I totally rewrote the TextGrid class so that it can:
- parse normal and short format TextGrids saved as text files in either UTF-8 or UTF-16 encoding
- does not load intervals with empty labels (by default, can be loaded with an optional `keep_empty` flag)
- can convert either all interval tiers to sequences (what you get by default when you call `to_seq` or `to_annot` with no `tier` argument) or a single specific tier (by supplying a `tier` argument)
as described here: https://github.com/vocalpy/crowsetta/issues/241

To make the it possible to convert all interval tiers, I also added a new feature for `crowsetta.Annotation`: the `seq` attribute can now be a list of `crowsetta.Sequence`s. This resolved a longstanding issue (https://github.com/vocalpy/crowsetta/issues/42).

Replying to note how I addressed comments:
> Maybe [the TextGrid format description](https://crowsetta.readthedocs.io/en/latest/formats/seq/textgrid.html#textgrid) could contain a short paragraph on how the format maps to crowsetta's data model?
I agree this needs to be clear, but would prefer to keep any description of how the format maps to crowsetta's data model in the docstrings with the code, in case it changes. I have added language saying basically "please read the docstring for further detail".
In https://github.com/vocalpy/crowsetta/pull/245

> > I found that, just like the code I have vendored now, textgridtools also raises a ValueError when opening most of the files from [this dataset](https://osf.io/r6paq/) because of overlapping tiers.
>  I think it would be fine to refuse these things in crowsetta
The new TextGrid class raises an error for overlapping intervals

> Since you can just read one tier, is there an easy way of merging Annotations?

You can now read more than one tier. A single Annotation can have multiple Sequences.

> In my experience, empty TextGrid intervals are almost always the equivalent of "no annotation" (given the way IntervalTiers work, as mentioned above). So I would suggest that maybe you could remove those by default, or at least provide an easy way for users to not do this manually.

The empty intervals are now removed by default.

> [even though it would seem to be supported by textgrid.py](https://github.com/kylebgorman/textgrid/tree/19d3ea2f765a435bb22a529602532752361167c9/tests/data), it can't read the unicode-containing short-format TextGrids I fed it

This is one of the main reasons I rewrote from scratch.

> But it might be good to note in the documentation that not all short TextGrids can be read?

Both short and full formats should now be parsed and the text encoding should be handled correctly.
I'll spare you the implementation details but the "Notes" section in the docstring points to other implementations I looked at.

> Binary-format TextGrid files cannot be read either, but I don't think this is very important.

I did find some implementations that handle the binary format but did not implement at this time, made an issue to add later

Thank you for providing the example TextGrid files.
I have included these in the test data for crowsetta and cited Parselmouth in a README there -- I note you use this for your own tests, I assume you exported each from Praat in the different formats?

> But maybe just specify that only text-based formats are supported.

Now noted in the docstring, along with a link to the issue about adding ability to load binary format.

> Mostly out of curiosity: my Segment.onset_sample and Segment.offset_sample are always `None. Is there a way to specify an audio file to get these filled in when reading a TextGrid? Or is this just not applicable in this case?

> Potentially, when converting a TextGrid to Annotation, it's nice if the user can specify a tier name instead of just the index. But then again, nothing urgent; more a thought of something useful.

You can now specify tiers by name or index.
E.g., either `textgrid[2]` or `textgrid['Phones']` will work.
See examples here:
https://crowsetta.readthedocs.io/en/latest/api/generated/crowsetta.formats.seq.textgrid.textgrid.TextGrid.html#crowsetta.formats.seq.textgrid.textgrid.TextGrid

> I don't think crowsetta means to do provide this, but I couldn't find a way to write TextGrids, right? So I don't need to have a look at that?

Crowsetta does mean to provide this but it doesn't right now.
https://github.com/vocalpy/crowsetta/issues/246

> In general, throughout the docs and especially the API reference, things like crowsetta.Annotation are not clickable.

I *totally* did not know you could do that in docs. Thanks so much for pointing this out. I have fixed it as much as possible.

Surprisingly it doesn't work for pandas. Seems to be an open issue related to it: https://github.com/pandas-dev/pandas/issues/38397

> The Tutorial page of the docs has an execution error somewhere

Fixed, thank you!
https://github.com/vocalpy/crowsetta/commit/413ea13bd38324a27a0ff2d82981213658e7ae32

> On the "How to use crowsetta to remove unlabeled intervals from TextGrid annotations" page, there's also something weird happening to the output during the download. Maybe there's a way of suppressing the output of that cell?

Suppressed
https://github.com/vocalpy/crowsetta/commit/0fdea17ee6ea93a72f39a9421bbce569d3c6a6ab
but now there are other issues with this tutorial.
I plan to rewrite totally now that empty intervals are done by default--I think it's still a good example of a full analysis using crowsetta.
https://github.com/vocalpy/crowsetta/issues/244

But this will have to wait until after today!
