### 2023-03-28 #68 @cmarmo
----------------------------------------------
🎉 [crowsetta](https://github.com/vocalpy/crowsetta) has been approved by pyOpenSci! Thank you @NickleDave  for submitting crowsetta and many thanks to @rhine3 and @shaupert  for reviewing this package, and @YannickJadoul for providing expertise on TextGrid format. 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of crowsetta. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.
- [ ] Add crowsetta to the pyOpenSci website. @NickleDave , please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors.
- [ ] @rhine3 @shaupert @YannickJadoul if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!

- [ ] When the JOSS review is complete, add a comment to your review in the pyOpenSci software-review repo that it has been approved by JOSS.
