### 2023-03-28 #68 @YannickJadoul
> Just merged a final PR that addresses feedback from @YannickJadoul along with some stray commits I made to main that I'll link here:

Wow, I'm impressed how you went above and beyond to address all those picked nits, @NickleDave!
(I do hope/think this will save you a bunch of time debugging Praat/TextGrid-related future issues from all those users :-) )

> I have included these in the test data for crowsetta and cited Parselmouth in a README there -- I note you use this for your own tests, I assume you exported each from Praat in the different formats?

To answer your final question: yes, I created the TextGrid myself to go along with [the fragment](https://github.com/YannickJadoul/Parselmouth/blob/dc14dc6d2e8733afabc9541abeeb603812fc75e6/tests/data/the_north_wind_and_the_sun.wav)  of the [audio file I've been using during developments and tests](https://en.wikipedia.org/wiki/File:Recording_of_speaker_of_British_English_(Received_Pronunciation).ogg) (and checked with people more knowledgeable about linguistics if it's +- correct ;-) ). Feel free to just use it without any need to reference Parselmouth!
And for the full record: yes, I converted [my original file](https://github.com/YannickJadoul/Parselmouth/blob/dc14dc6d2e8733afabc9541abeeb603812fc75e6/tests/data/the_north_wind_and_the_sun.TextGrid) to all other formats with Praat 6.2.23.
