### 2023-03-29 45ad64d
DOC: Update CHANGELOG with info on #243 [skip ci]
