### 2023-03-29 94c7446
CI: Drop testing on Python 3.8
