+++ README.md
@@ -192,7 +192,7 @@ Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/d
 <table>
   <tbody>
     <tr>
-      <td align="center" valign="top" width="14.28%"><a href="http://tessarhinehart.com"><img src="https://avatars.githubusercontent.com/u/27306428?v=4?s=100" width="100px;" alt="Tessa Rhinehart"/><br /><sub><b>Tessa Rhinehart</b></sub></a><br /><a href="https://github.com/vocalpy/crowsetta/commits?author=rhine3" title="Documentation">📖</a> <a href="https://github.com/vocalpy/crowsetta/issues?q=author%3Arhine3" title="Bug reports">🐛</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="http://tessarhinehart.com"><img src="https://avatars.githubusercontent.com/u/27306428?v=4?s=100" width="100px;" alt="Tessa Rhinehart"/><br /><sub><b>Tessa Rhinehart</b></sub></a><br /><a href="https://github.com/vocalpy/crowsetta/commits?author=rhine3" title="Documentation">📖</a> <a href="https://github.com/vocalpy/crowsetta/issues?q=author%3Arhine3" title="Bug reports">🐛</a> <a href="#userTesting-rhine3" title="User Testing">📓</a> <a href="#ideas-rhine3" title="Ideas, Planning, & Feedback">🤔</a></td>
       <td align="center" valign="top" width="14.28%"><a href="https://github.com/shaupert"><img src="https://avatars.githubusercontent.com/u/43136040?v=4?s=100" width="100px;" alt="Sylvain HAUPERT"/><br /><sub><b>Sylvain HAUPERT</b></sub></a><br /><a href="https://github.com/vocalpy/crowsetta/commits?author=shaupert" title="Code">💻</a> <a href="#ideas-shaupert" title="Ideas, Planning, & Feedback">🤔</a></td>
       <td align="center" valign="top" width="14.28%"><a href="https://github.com/YannickJadoul"><img src="https://avatars.githubusercontent.com/u/8025744?v=4?s=100" width="100px;" alt="Yannick Jadoul"/><br /><sub><b>Yannick Jadoul</b></sub></a><br /><a href="#ideas-YannickJadoul" title="Ideas, Planning, & Feedback">🤔</a> <a href="https://github.com/vocalpy/crowsetta/issues?q=author%3AYannickJadoul" title="Bug reports">🐛</a> <a href="https://github.com/vocalpy/crowsetta/commits?author=YannickJadoul" title="Documentation">📖</a> <a href="#userTesting-YannickJadoul" title="User Testing">📓</a></td>
     </tr>
