+++ pyproject.toml
@@ -21,7 +21,7 @@ dependencies = [
     "scipy >=1.7.0",
     "SoundFile >=0.12.1",
 ]
-version = "5.0.0rc2"
+version = "5.0.0"
 readme = "README.md"
 license = {file = "LICENSE"}
 classifiers = [
