### 2023-03-29 991b375
DOC: Add pyOpenSci badge to README [skip ci]
