### 2023-03-29 #68 @NickleDave
Hi @cmarmo just updating here that I made a release, here's the DOI: https://zenodo.org/record/7781587.
Also made the PR adding crowsetta to packages.yml. Will proceed with JOSS submission now.
