### 2023-03-29 #68 @cmarmo
@rhine3 @shaupert just to let you know that you should have both received an invitation to join the pyOpenSci Slack, if you are interested to jon. Thanks!
