### 2023-03-29 #68 @lwasser
wow - everyone!! this is awesome. Congratulations @NickleDave !! and @cmarmo for a successful review! And so many thanks @shaupert  @YannickJadoul @rhine3 for your time in reviewing here! David, let us know when this goes to JOSS! in the meantime i'll post about crowsetta on social.

We'd also love a blog about the package if you have time. 😄 This blog allows us to better promote your package / show what your package can do!
