### 2023-03-30 e2710e7
DOC: Rewrite remove-silent-labels-textgrid.md, fix #244
- Rename/rewrite -> doc/howto/munging-annotations-textgrid.md
- Add 4 TextGrid files in doc/data/r6paq/
