+++ doc/howto/munging-annotations-textgrid.md
@@ -9,11 +9,13 @@ kernelspec:
   display_name: Python 3 (ipykernel)
   language: python
   name: python3
+mystnb:
+  execution_timeout: 300
 ---
 
-(howto-remove-silent-labels-textgrid)=
+(munging-annotations-textgrid)=
 
-# How to use crowsetta to remove unlabeled intervals from TextGrid annotations
+# Munging annotations from Praat TextGrid files with crowsetta to fit a first-order Markov model of birdsong syntax
 
 A common need when working with annotations 
 for animal vocalizations 
@@ -21,39 +23,12 @@ is to transform them to perform a certain analysis.
 This is a sort of [data munging](https://en.wikipedia.org/wiki/Data_wrangling).
 In this vignette, we provide a simple example of munging annotations to build a model of birdsong syntax.
 
-We will work with a dataset 
-of Bengalese finch song annotated with the 
-[Praat .TextGrid format](https://www.fon.hum.uva.nl/praat/manual/TextGrid_file_formats.html).
+We will work with a 
+[dataset of Bengalese finch song](https://nickledave.github.io/bfsongrepo/)
+annotated in a format used by ["evsonganaly"](https://github.com/NickleDave/evfuncs),
+a MATLAB GUI for annotating song.
 We want to build a matrix of the transition probabilities between syllables in the song, 
 that is, a first-order [Markov model](https://en.wikipedia.org/wiki/Markov_chain).
-But before we can do that, 
-we need to clean the data.
-
-Specifically, we need to remove the labels that Praat gives 
-to all of the silent periods between syllables when it saves a TextGrid file.
-In the terminology of Praat, 
-we need to remove the *intervals* 
-that have empty labels.
-[Praat allows annotators to create multiple *tiers*](https://www.fon.hum.uva.nl/praat/manual/Intro_7__Annotation.html), or
-or levels, of labeling 
-E.g., if we were annotating speech, 
-we could have a tier each for phonemes, syllables, 
-words, and sentences.
-In this example we will work with interval tiers, 
-that consists of a set of *intervals* as the name suggests.
-(There are also point tiers where annotators label single time points.)
-Each interval has a start time, stop time, and label,
-all created by an annotator working with the Praat GUI.
-When the annotator does not assign a label to some interval, 
-then the corresponding TextGrid file will contain an interval with an empty label. 
-Commonly this happens for silent periods between intervals of interest.
-In our case, it happens for every brief (50-100ms) silent gap 
-between the syllables in a Bengalese finch's song.
-These lables become `None` when we load them into crowsetta.
-We don't want to include these intervals as states in our Markov model, 
-because we are only interested in transitions between syllables.
-So, to compute our transition matrix, 
-we need to remove all these intervals with `None` labels.
 
 +++
 
@@ -68,25 +43,22 @@ after installing Jupyter lab and jupytext.
 
 +++
 
-
-
-
 ## Workflow
 
 Here's the steps we'll follow:  
 
 0. Write code for analysis
-1. Download data from Open Science Framework
-2. Load data with crowsetta
-3. Remove `None` labels
-4. Compute transition matrix
+1. Load .not.mat files saved by evsonganaly with crowsetta, and convert to crowsetta Annotations with Sequences
+2. Munge data
+3. Compute transition matrices
 
 For this vignette, we use annotations from the 
-[Bengalese finch song dataset](https://osf.io/r6paq/), 
-by Tachibana and Morita 2021 [^1], adapted under 
+[Bengalese finch song repository](https://nickledave.github.io/bfsongrepo), 
+by Nicholson, Queen, Sober 2017 [^1], adapted under 
 [CC BY 4.0 License](https://creativecommons.org/licenses/by/4.0/).
 
[…]
