+++ pyproject.toml
@@ -14,7 +14,6 @@ dependencies = [
     "attrs >=19.3.0",
     "evfuncs >=0.3.5",
     "birdsong-recognition-dataset >=0.3.2",
-    "importlib-resources >=5.7.1",
     "numpy >=1.21.0",
     "pandas >= 1.3.5",
     "pandera >= 0.9.0",
