### 2023-03-31 b3d33bc
DEV: Make nox session test run Python 3.9, 3.10, 3.11
