+++ noxfile.py
@@ -66,7 +66,14 @@ def lint(session):
     session.run("flake8", "./src", "--max-line-length", "120", "--exclude", "./src/crowsetta/_vendor")
 
 
-@nox.session
+TEST_PYTHONS = [
+    "3.9",
+    "3.10",
+    "3.11"
+]
+
+
+@nox.session(python=TEST_PYTHONS)
 def test(session) -> None:
     """
     Run the unit and regular tests.
