### 2023-03-31 21a2e51
CI: Run lint session on Python 3.10
