+++ .github/workflows/lint.yml
@@ -13,7 +13,7 @@ jobs:
           src: "./src"
       - uses: actions/setup-python@v2
         with:
-          python-version: 3.8
+          python-version: "3.10"
       - name: install libsndfile1 on ubuntu
         if: matrix.os == 'ubuntu-latest'
         run: sudo apt install libsndfile1
