### 2023-04-01 e68c699
DOC: Finish munging howto fixup, other clean-ups
- Remove doc/data/r6paq/b17_concat_dtw.TextGrid
  because it's 11 MB and takes forever to load with
  the new TextGrid class
- Fix toctree in doc/howto.md
- Fix links in howto/convert-simple-seq.md
- Fix download links in doc/howto/munging-annotations-textgrid.md
