### 2023-04-04 7f14eed
DOC: Fix Textgrid snippet in doc/index.md [skip ci]
