### 2023-04-13 #68 @NickleDave
@cmarmo accepted into JOSS! 🎉
https://github.com/openjournals/joss-reviews/issues/5338#issuecomment-1506566173
https://joss.theoj.org/papers/10.21105/joss.05338
