### 2023-04-13 cfb2cb8
DOC: Add JOSS badge to README [skip ci]
